% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ACT1_2A}
\alias{ACT1_2A}
\title{ACT1, figure 2A}
\format{
A data frame of 637 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (amrubicin, topotecan) \cr
}
}
\source{
von Pawel J, Jotte R, Spigel DR, et al. Randomized phase III trial of
amrubicin versus topotecan as second-line treatment for patients with
small-cell lung cancer. J Clin Oncol 2014; 32: 4012–9.
}
\usage{
ACT1_2A
}
\description{
Kaplan-Meier digitized data from ACT1, figure 2A (PMID 25385727). A reported sample size of 637 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(ACT1_2A)

kmplot(ACT1_2A)
}
\keyword{datasets}
