% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ACT2_2C}
\alias{ACT2_2C}
\title{ACT2, figure 2C}
\format{
A data frame of 71 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (wt_b, wt_be) \cr
}
}
\source{
Hagman H, Frödin J-E, Berglund Å, et al. A randomized study of
KRAS-guided maintenance therapy with bevacizumab, erlotinib or
metronomic capecitabine after first-line induction treatment of
metastatic colorectal cancer: the Nordic ACT2 trial. Ann Oncol 2016;
27: 140–7.
}
\usage{
ACT2_2C
}
\description{
Kaplan-Meier digitized data from ACT2, figure 2C (PMID 26483047). A reported sample size of 233 for a primary endpoint of PFS-3mo in colorectal cancer.
}
\examples{
summary(ACT2_2C)

kmplot(ACT2_2C)
}
\keyword{datasets}
