% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ACTSCC_2A}
\alias{ACTSCC_2A}
\title{ACTSCC, figure 2A}
\format{
A data frame of 1,518 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (s_1, uft_lv) \cr
}
}
\source{
Yoshida M, Ishiguro M, Ikejiri K, et al. S-1 as adjuvant chemotherapy
for stage III colon cancer: a randomized phase III study (ACTS-CC
trial). Ann Oncol 2014; 25: 1743–9.
}
\usage{
ACTSCC_2A
}
\description{
Kaplan-Meier digitized data from ACTSCC, figure 2A (PMID 24942277). A reported sample size of 1,518 for a primary endpoint of DFS_3yr in colorectal cancer.
}
\examples{
summary(ACTSCC_2A)

kmplot(ACTSCC_2A)
}
\keyword{datasets}
