% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AIO0207(b)_3B}
\alias{AIO0207(b)_3B}
\title{AIO0207(b), figure 3B}
\format{
A data frame of 300 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (5fu_bev, bev) \cr
}
}
\source{
Hegewisch-Becker S, Graeven U, Lerchenmüller CA, et al. Maintenance
strategies after first-line oxaliplatin plus fluoropyrimidine plus
bevacizumab for patients with metastatic colorectal cancer (AIO
0207): a randomised, non-inferiority, open-label, phase 3 trial.
Lancet Oncol 2015; 16: 1355–69.
}
\usage{
`AIO0207(b)_3B`
}
\description{
Kaplan-Meier digitized data from AIO0207(b), figure 3B (PMID 26361971). A reported sample size of 472 for a primary endpoint of time_tx_failure in colorectal cancer.
}
\examples{
summary(`AIO0207(b)_3B`)

kmplot(`AIO0207(b)_3B`)
}
\keyword{datasets}
