% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ALTTO(b)_2B}
\alias{ALTTO(b)_2B}
\title{ALTTO(b), figure 2B}
\format{
A data frame of 4,188 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (t, t_then_l) \cr
}
}
\source{
Piccart-Gebhart M, Holmes E, Baselga J, et al. Adjuvant Lapatinib and
Trastuzumab for Early Human Epidermal Growth Factor Receptor
2-Positive Breast Cancer: Results From the Randomized Phase III
Adjuvant Lapatinib and/or Trastuzumab Treatment Optimization Trial. J
Clin Oncol 2016; 34: 1034–42.
}
\usage{
`ALTTO(b)_2B`
}
\description{
Kaplan-Meier digitized data from ALTTO(b), figure 2B (PMID 26598744). A reported sample size of 8,381 for a primary endpoint of DFS in breast cancer.
}
\examples{
summary(`ALTTO(b)_2B`)

kmplot(`ALTTO(b)_2B`)
}
\keyword{datasets}
