% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMAROS_2B}
\alias{AMAROS_2B}
\title{AMAROS, figure 2B}
\format{
A data frame of 1,425 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (dissection, radiotherapy) \cr
}
}
\source{
Donker M, van Tienhoven G, Straver ME, et al. Radiotherapy or surgery
of the axilla after a positive sentinel node in breast cancer (EORTC
10981-22023 AMAROS): a randomised, multicentre, open-label, phase 3
non-inferiority trial. Lancet Oncol 2014; 15: 1303–10.
}
\usage{
AMAROS_2B
}
\description{
Kaplan-Meier digitized data from AMAROS, figure 2B (PMID 25439688). A reported sample size of 4,823 for a primary endpoint of AxRecurrence-5yr in breast cancer.
}
\examples{
summary(AMAROS_2B)

kmplot(AMAROS_2B)
}
\keyword{datasets}
