% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASPECCT_3A}
\alias{ASPECCT_3A}
\title{ASPECCT, figure 3A}
\format{
A data frame of 999 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (cetuximab, panitumumab) \cr
}
}
\source{
Price TJ, Peeters M, Kim TW, et al. Panitumumab versus cetuximab in
patients with chemotherapy-refractory wild-type KRAS exon 2
metastatic colorectal cancer (ASPECCT): a randomised, multicentre,
open-label, non-inferiority phase 3 study. Lancet Oncol 2014; 15:
569–79.
}
\usage{
ASPECCT_3A
}
\description{
Kaplan-Meier digitized data from ASPECCT, figure 3A (PMID 24739896). A reported sample size of 1,010 for a primary endpoint of OS in colorectal cancer.
}
\examples{
summary(ASPECCT_3A)

kmplot(ASPECCT_3A)
}
\keyword{datasets}
