% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ATTENTION_2A}
\alias{ATTENTION_2A}
\title{ATTENTION, figure 2A}
\format{
A data frame of 307 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (placebo, tivantinib) \cr
}
}
\source{
Yoshioka H, Azuma K, Yamamoto N, et al. A randomized, double-blind,
placebo-controlled, phase III trial of erlotinib with or without a
c-Met inhibitor tivantinib (ARQ 197) in Asian patients with
previously treated stage IIIB/IV nonsquamous nonsmall-cell lung
cancer harboring wild-type epidermal growth factor receptor
(ATTENTION study). Ann Oncol 2015; 26: 2066–72.
}
\usage{
ATTENTION_2A
}
\description{
Kaplan-Meier digitized data from ATTENTION, figure 2A (PMID 26153496). A reported sample size of 460 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(ATTENTION_2A)

kmplot(ATTENTION_2A)
}
\keyword{datasets}
