% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AURA3_1A}
\alias{AURA3_1A}
\title{AURA3, figure 1A}
\format{
A data frame of 419 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (osimertinib, platinum-pemetrexed) \cr
}
}
\source{
Mok TS, Wu Y-L, Ahn M-J, et al. Osimertinib or Platinum-Pemetrexed in
EGFR T790M-Positive Lung Cancer. N Engl J Med 2017; 376: 629–40.
}
\usage{
AURA3_1A
}
\description{
Kaplan-Meier digitized data from AURA3, figure 1A (PMID 27959700). A reported sample size of 419 for a primary endpoint of PFS in lung cancer.
}
\examples{
summary(AURA3_1A)

kmplot(AURA3_1A)
}
\keyword{datasets}
