% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AZURE_2A}
\alias{AZURE_2A}
\title{AZURE, figure 2A}
\format{
A data frame of 3,359 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (control, zoledronic_acid) \cr
}
}
\source{
Coleman R, Cameron D, Dodwell D, et al. Adjuvant zoledronic acid in
patients with early breast cancer: final efficacy analysis of the
AZURE (BIG 01/04) randomised open-label phase 3 trial. Lancet Oncol
2014; 15: 997–1006.
}
\usage{
AZURE_2A
}
\description{
Kaplan-Meier digitized data from AZURE, figure 2A (PMID 25035292). A reported sample size of 3,360 for a primary endpoint of DFS in breast cancer.
}
\examples{
summary(AZURE_2A)

kmplot(AZURE_2A)
}
\keyword{datasets}
