% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BEACON_2B}
\alias{BEACON_2B}
\title{BEACON, figure 2B}
\format{
A data frame of 852 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (control, etirinotecan_pegol) \cr
}
}
\source{
Perez EA, Awada A, O’Shaughnessy J, et al. Etirinotecan pegol
(NKTR-102) versus treatment of physician’s choice in women with
advanced breast cancer previously treated with an anthracycline, a
taxane, and capecitabine (BEACON): a randomised, open-label,
multicentre, phase 3 trial. Lancet Oncol 2015; 16: 1556–68.
}
\usage{
BEACON_2B
}
\description{
Kaplan-Meier digitized data from BEACON, figure 2B (PMID 26482278). A reported sample size of 852 for a primary endpoint of OS in breast cancer.
}
\examples{
summary(BEACON_2B)

kmplot(BEACON_2B)
}
\keyword{datasets}
