% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BEBYP_1}
\alias{BEBYP_1}
\title{BEBYP, figure 1}
\format{
A data frame of 184 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (bev_ct, ct) \cr
}
}
\source{
Masi G, Salvatore L, Boni L, et al. Continuation or reintroduction of
bevacizumab beyond progression to first-line therapy in metastatic
colorectal cancer: final results of the randomized BEBYP trial. Ann
Oncol 2015; 26: 724–30.
}
\usage{
BEBYP_1
}
\description{
Kaplan-Meier digitized data from BEBYP, figure 1 (PMID 25600568). A reported sample size of 185 for a primary endpoint of PFS in colorectal cancer.
}
\examples{
summary(BEBYP_1)

kmplot(BEBYP_1)
}
\keyword{datasets}
