% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BEYOND_2A}
\alias{BEYOND_2A}
\title{BEYOND, figure 2A}
\format{
A data frame of 276 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (bev_cp, placebo_cp) \cr
}
}
\source{
Zhou C, Wu Y-L, Chen G, et al. BEYOND: A Randomized, Double-Blind,
Placebo-Controlled, Multicenter, Phase III Study of First-Line
Carboplatin/Paclitaxel Plus Bevacizumab or Placebo in Chinese
Patients With Advanced or Recurrent Nonsquamous Non-Small-Cell Lung
Cancer. J Clin Oncol 2015; 33: 2197–204.
}
\usage{
BEYOND_2A
}
\description{
Kaplan-Meier digitized data from BEYOND, figure 2A (PMID 26014294). A reported sample size of 276 for a primary endpoint of PFS in prostate cancer.
}
\examples{
summary(BEYOND_2A)

kmplot(BEYOND_2A)
}
\keyword{datasets}
