% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BOLERO1_2B}
\alias{BOLERO1_2B}
\title{BOLERO1, figure 2B}
\format{
A data frame of 719 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (everolimus, placebo) \cr
}
}
\source{
Hurvitz SA, Andre F, Jiang Z, et al. Combination of everolimus with
trastuzumab plus paclitaxel as first-line treatment for patients with
HER2-positive advanced breast cancer (BOLERO-1): a phase 3,
randomised, double-blind, multicentre trial. Lancet Oncol 2015; 16:
816–29.
}
\usage{
BOLERO1_2B
}
\description{
Kaplan-Meier digitized data from BOLERO1, figure 2B (PMID 26092818). A reported sample size of 719 for a primary endpoint of PFS(overall & HR- population) in breast cancer.
}
\examples{
summary(BOLERO1_2B)

kmplot(BOLERO1_2B)
}
\keyword{datasets}
