% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BOLERO2_1}
\alias{BOLERO2_1}
\title{BOLERO2, figure 1}
\format{
A data frame of 724 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (eve_exe, pbo_exe) \cr
}
}
\source{
Piccart M, Hortobagyi GN, Campone M, et al. Everolimus plus
exemestane for hormone-receptor-positive, human epidermal growth
factor receptor-2-negative advanced breast cancer: overall survival
results from BOLERO-2†. Ann Oncol 2014; 25: 2357–62.
}
\usage{
BOLERO2_1
}
\description{
Kaplan-Meier digitized data from BOLERO2, figure 1 (PMID 25231953). A reported sample size of 720 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(BOLERO2_1)

kmplot(BOLERO2_1)
}
\keyword{datasets}
