% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BOLERO3_2}
\alias{BOLERO3_2}
\title{BOLERO3, figure 2}
\format{
A data frame of 569 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in weeks) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (eve_exe, placebo) \cr
}
}
\source{
André F, O’Regan R, Ozguroglu M, et al. Everolimus for women with
trastuzumab-resistant, HER2-positive, advanced breast cancer
(BOLERO-3): a randomised, double-blind, placebo-controlled phase 3
trial. Lancet Oncol 2014; 15: 580–91.
}
\usage{
BOLERO3_2
}
\description{
Kaplan-Meier digitized data from BOLERO3, figure 2 (PMID 24742739). A reported sample size of 569 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(BOLERO3_2)

kmplot(BOLERO3_2)
}
\keyword{datasets}
