% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BREC_2A}
\alias{BREC_2A}
\title{BREC, figure 2A}
\format{
A data frame of 277 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (control, experimental) \cr
}
}
\source{
Moran T, Wei J, Cobo M, et al. Two biomarker-directed randomized
trials in European and Chinese patients with nonsmall-cell lung
cancer: the BRCA1-RAP80 Expression Customization (BREC) studies. Ann
Oncol 2014; 25: 2147–55.
}
\usage{
BREC_2A
}
\description{
Kaplan-Meier digitized data from BREC, figure 2A (PMID 25164908). A reported sample size of 279 for a primary endpoint of PFS in lung cancer.
}
\examples{
summary(BREC_2A)

kmplot(BREC_2A)
}
\keyword{datasets}
