% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CA097375_2}
\alias{CA097375_2}
\title{CA097375, figure 2}
\format{
A data frame of 499 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (a, b) \cr
}
}
\source{
Love RR, Laudico AV, Van Dinh N, et al. Timing of adjuvant surgical
oophorectomy in the menstrual cycle and disease-free and overall
survival in premenopausal women with operable breast cancer. J Natl
Cancer Inst 2015; 107: djv064.
}
\usage{
CA097375_2
}
\description{
Kaplan-Meier digitized data from CA097375, figure 2 (PMID 25794890). A reported sample size of 740 for a primary endpoint of DFS in breast cancer.
}
\examples{
summary(CA097375_2)

kmplot(CA097375_2)
}
\keyword{datasets}
