% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CA184043_2A}
\alias{CA184043_2A}
\title{CA184043, figure 2A}
\format{
A data frame of 799 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (ipilimumab, placebo) \cr
}
}
\source{
Kwon ED, Drake CG, Scher HI, et al. Ipilimumab versus placebo after
radiotherapy in patients with metastatic castration-resistant
prostate cancer that had progressed after docetaxel chemotherapy
(CA184-043): a multicentre, randomised, double-blind, phase 3 trial.
Lancet Oncol 2014; 15: 700–12.
}
\usage{
CA184043_2A
}
\description{
Kaplan-Meier digitized data from CA184043, figure 2A (PMID 24831977). A reported sample size of 799 for a primary endpoint of OS in prostate cancer.
}
\examples{
summary(CA184043_2A)

kmplot(CA184043_2A)
}
\keyword{datasets}
