% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CAIRO3_2B}
\alias{CAIRO3_2B}
\title{CAIRO3, figure 2B}
\format{
A data frame of 557 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (maintenance, observation) \cr
}
}
\source{
Simkens LHJ, van Tinteren H, May A, et al. Maintenance treatment with
capecitabine and bevacizumab in metastatic colorectal cancer
(CAIRO3): a phase 3 randomised controlled trial of the Dutch
Colorectal Cancer Group. Lancet 2015; 385: 1843–52.
}
\usage{
CAIRO3_2B
}
\description{
Kaplan-Meier digitized data from CAIRO3, figure 2B (PMID 25862517). A reported sample size of 558 for a primary endpoint of PFS2 in colorectal cancer.
}
\examples{
summary(CAIRO3_2B)

kmplot(CAIRO3_2B)
}
\keyword{datasets}
