% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CALGB40502_2C}
\alias{CALGB40502_2C}
\title{CALGB40502, figure 2C}
\format{
A data frame of 487 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (ixabepilone, p1) \cr
}
}
\source{
Rugo HS, Barry WT, Moreno-Aspitia A, et al. Randomized Phase III
Trial of Paclitaxel Once Per Week Compared With Nanoparticle
Albumin-Bound Nab-Paclitaxel Once Per Week or Ixabepilone With
Bevacizumab As First-Line Chemotherapy for Locally Recurrent or
Metastatic Breast Cancer: CALGB 40502/NCCTG N063H (Alliance). J Clin
Oncol 2015; 33: 2361–9.
}
\usage{
CALGB40502_2C
}
\description{
Kaplan-Meier digitized data from CALGB40502, figure 2C (PMID 26056183). A reported sample size of 799 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(CALGB40502_2C)

kmplot(CALGB40502_2C)
}
\keyword{datasets}
