% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CALGB40503_2A}
\alias{CALGB40503_2A}
\title{CALGB40503, figure 2A}
\format{
A data frame of 343 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (l, l_b) \cr
}
}
\source{
Dickler MN, Barry WT, Cirrincione CT, et al. Phase III Trial
Evaluating Letrozole As First-Line Endocrine Therapy With or Without
Bevacizumab for the Treatment of Postmenopausal Women With Hormone
Receptor-Positive Advanced-Stage Breast Cancer: CALGB 40503
(Alliance). J Clin Oncol 2016; 34: 2602–9.
}
\usage{
CALGB40503_2A
}
\description{
Kaplan-Meier digitized data from CALGB40503, figure 2A (PMID 27138575). A reported sample size of 350 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(CALGB40503_2A)

kmplot(CALGB40503_2A)
}
\keyword{datasets}
