% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CALGB90202_2B}
\alias{CALGB90202_2B}
\title{CALGB90202, figure 2B}
\format{
A data frame of 645 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (placebo, za) \cr
}
}
\source{
Smith MR, Halabi S, Ryan CJ, et al. Randomized controlled trial of
early zoledronic acid in men with castration-sensitive prostate
cancer and bone metastases: results of CALGB 90202 (alliance). J Clin
Oncol 2014; 32: 1143–50.
}
\usage{
CALGB90202_2B
}
\description{
Kaplan-Meier digitized data from CALGB90202, figure 2B (PMID 24590644). A reported sample size of 645 for a primary endpoint of time_SRE in prostate cancer.
}
\examples{
summary(CALGB90202_2B)

kmplot(CALGB90202_2B)
}
\keyword{datasets}
