% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CBCSG006_2B}
\alias{CBCSG006_2B}
\title{CBCSG006, figure 2B}
\format{
A data frame of 236 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (cis_gemcitabine, paclitaxel_gemcitabine) \cr
}
}
\source{
Hu X-C, Zhang J, Xu B-H, et al. Cisplatin plus gemcitabine versus
paclitaxel plus gemcitabine as first-line therapy for metastatic
triple-negative breast cancer (CBCSG006): a randomised, open-label,
multicentre, phase 3 trial. Lancet Oncol 2015; 16: 436–46.
}
\usage{
CBCSG006_2B
}
\description{
Kaplan-Meier digitized data from CBCSG006, figure 2B (PMID 25795409). A reported sample size of 240 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(CBCSG006_2B)

kmplot(CBCSG006_2B)
}
\keyword{datasets}
