% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Checkmate017_1}
\alias{Checkmate017_1}
\title{Checkmate017, figure 1}
\format{
A data frame of 272 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (d1, nivolumab) \cr
}
}
\source{
Brahmer J, Reckamp KL, Baas P, et al. Nivolumab versus Docetaxel in
Advanced Squamous-Cell Non-Small-Cell Lung Cancer. N Engl J Med 2015;
373: 123–35.
}
\usage{
Checkmate017_1
}
\description{
Kaplan-Meier digitized data from Checkmate017, figure 1 (PMID 26028407). A reported sample size of 272 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(Checkmate017_1)

kmplot(Checkmate017_1)
}
\keyword{datasets}
