% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Checkmate057_1C}
\alias{Checkmate057_1C}
\title{Checkmate057, figure 1C}
\format{
A data frame of 582 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (d1, nivolumab) \cr
}
}
\source{
Borghaei H, Paz-Ares L, Horn L, et al. Nivolumab versus Docetaxel in
Advanced Nonsquamous Non-Small-Cell Lung Cancer. N Engl J Med 2015;
373: 1627–39.
}
\usage{
Checkmate057_1C
}
\description{
Kaplan-Meier digitized data from Checkmate057, figure 1C (PMID 26412456). A reported sample size of 582 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(Checkmate057_1C)

kmplot(Checkmate057_1C)
}
\keyword{datasets}
