% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CHHiP(a)_2A}
\alias{CHHiP(a)_2A}
\title{CHHiP(a), figure 2A}
\format{
A data frame of 2,139 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab RFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (60gy, 74gy) \cr
}
}
\source{
Dearnaley D, Syndikus I, Mossop H, et al. Conventional versus
hypofractionated high-dose intensity-modulated radiotherapy for
prostate cancer: 5-year outcomes of the randomised, non-inferiority,
phase 3 CHHiP trial. Lancet Oncol 2016; 17: 1047–60.
}
\usage{
`CHHiP(a)_2A`
}
\description{
Kaplan-Meier digitized data from CHHiP(a), figure 2A (PMID 27339115). A reported sample size of 3,216 for a primary endpoint of bRFS in prostate cancer.
}
\examples{
summary(`CHHiP(a)_2A`)

kmplot(`CHHiP(a)_2A`)
}
\keyword{datasets}
