% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CHHiP(b)_2B}
\alias{CHHiP(b)_2B}
\title{CHHiP(b), figure 2B}
\format{
A data frame of 2,142 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (57gy, 74gy) \cr
}
}
\source{
Dearnaley D, Syndikus I, Mossop H, et al. Conventional versus
hypofractionated high-dose intensity-modulated radiotherapy for
prostate cancer: 5-year outcomes of the randomised, non-inferiority,
phase 3 CHHiP trial. Lancet Oncol 2016; 17: 1047–60.
}
\usage{
`CHHiP(b)_2B`
}
\description{
Kaplan-Meier digitized data from CHHiP(b), figure 2B (PMID 27339115). A reported sample size of 3,216 for a primary endpoint of bRFS in prostate cancer.
}
\examples{
summary(`CHHiP(b)_2B`)

kmplot(`CHHiP(b)_2B`)
}
\keyword{datasets}
