% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Chronicle_2A}
\alias{Chronicle_2A}
\title{Chronicle, figure 2A}
\format{
A data frame of 113 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (capecitabine_oxaliplatin, follow_up_only) \cr
}
}
\source{
Glynne-Jones R, Counsell N, Quirke P, et al. Chronicle: results of a
randomised phase III trial in locally advanced rectal cancer after
neoadjuvant chemoradiation randomising postoperative adjuvant
capecitabine plus oxaliplatin (XELOX) versus control. Ann Oncol 2014;
25: 1356–62.
}
\usage{
Chronicle_2A
}
\description{
Kaplan-Meier digitized data from Chronicle, figure 2A (PMID 24718885). A reported sample size of 113 for a primary endpoint of DFS in colorectal cancer.
}
\examples{
summary(Chronicle_2A)

kmplot(Chronicle_2A)
}
\keyword{datasets}
