% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CLEOPATRA_2A}
\alias{CLEOPATRA_2A}
\title{CLEOPATRA, figure 2A}
\format{
A data frame of 808 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (control, pertuzumab) \cr
}
}
\source{
Swain SM, Baselga J, Kim S-B, et al. Pertuzumab, trastuzumab, and
docetaxel in HER2-positive metastatic breast cancer. N Engl J Med
2015; 372: 724–34.
}
\usage{
CLEOPATRA_2A
}
\description{
Kaplan-Meier digitized data from CLEOPATRA, figure 2A (PMID 25693012). A reported sample size of 808 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(CLEOPATRA_2A)

kmplot(CLEOPATRA_2A)
}
\keyword{datasets}
