% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COMET1_2B}
\alias{COMET1_2B}
\title{COMET1, figure 2B}
\format{
A data frame of 1,028 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (cabozantinib, prednisone) \cr
}
}
\source{
Smith M, De Bono J, Sternberg C, et al. Phase III Study of
Cabozantinib in Previously Treated Metastatic Castration-Resistant
Prostate Cancer: COMET-1. J Clin Oncol 2016; 34: 3005–13.
}
\usage{
COMET1_2B
}
\description{
Kaplan-Meier digitized data from COMET1, figure 2B (PMID 27400947). A reported sample size of 1,028 for a primary endpoint of OS in prostate cancer.
}
\examples{
summary(COMET1_2B)

kmplot(COMET1_2B)
}
\keyword{datasets}
