% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONCEPT_1A}
\alias{CONCEPT_1A}
\title{CONCEPT, figure 1A}
\format{
A data frame of 139 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab TTF event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (co, io) \cr
}
}
\source{
Hochster HS, Grothey A, Hart L, et al. Improved time to treatment
failure with an intermittent oxaliplatin strategy: results of
CONcePT. Ann Oncol 2014; 25: 1172–8.
}
\usage{
CONCEPT_1A
}
\description{
Kaplan-Meier digitized data from CONCEPT, figure 1A (PMID 24608198). A reported sample size of 140 for a primary endpoint of TTF in colorectal cancer.
}
\examples{
summary(CONCEPT_1A)

kmplot(CONCEPT_1A)
}
\keyword{datasets}
