% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONCUR_2A}
\alias{CONCUR_2A}
\title{CONCUR, figure 2A}
\format{
A data frame of 204 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (placebo, regorafenib) \cr
}
}
\source{
Li J, Qin S, Xu R, et al. Regorafenib plus best supportive care
versus placebo plus best supportive care in Asian patients with
previously treated metastatic colorectal cancer (CONCUR): a
randomised, double-blind, placebo-controlled, phase 3 trial. Lancet
Oncol 2015; 16: 619–29.
}
\usage{
CONCUR_2A
}
\description{
Kaplan-Meier digitized data from CONCUR, figure 2A (PMID 25981818). A reported sample size of 243 for a primary endpoint of OS in colorectal cancer.
}
\examples{
summary(CONCUR_2A)

kmplot(CONCUR_2A)
}
\keyword{datasets}
