% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DART0105_2B}
\alias{DART0105_2B}
\title{DART0105, figure 2B}
\format{
A data frame of 347 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (ltad, stad) \cr
}
}
\source{
Zapatero A, Guerrero A, Maldonado X, et al. High-dose radiotherapy
with short-term or long-term androgen deprivation in localised
prostate cancer (DART01/05 GICOR): a randomised, controlled, phase 3
trial. Lancet Oncol 2015; 16: 320–7.
}
\usage{
DART0105_2B
}
\description{
Kaplan-Meier digitized data from DART0105, figure 2B (PMID 25702876). A reported sample size of 178 for a primary endpoint of bDFS in prostate cancer.
}
\examples{
summary(DART0105_2B)

kmplot(DART0105_2B)
}
\keyword{datasets}
