% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DELTA_2B}
\alias{DELTA_2B}
\title{DELTA, figure 2B}
\format{
A data frame of 301 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (d1, erlotinib) \cr
}
}
\source{
Kawaguchi T, Ando M, Asami K, et al. Randomized phase III trial of
erlotinib versus docetaxel as second- or third-line therapy in
patients with advanced non-small-cell lung cancer: Docetaxel and
Erlotinib Lung Cancer Trial (DELTA). J Clin Oncol 2014; 32: 1902–8.
}
\usage{
DELTA_2B
}
\description{
Kaplan-Meier digitized data from DELTA, figure 2B (PMID 24841974). A reported sample size of 301 for a primary endpoint of PFS in lung cancer.
}
\examples{
summary(DELTA_2B)

kmplot(DELTA_2B)
}
\keyword{datasets}
