% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{E1199(b)_2C}
\alias{E1199(b)_2C}
\title{E1199(b), figure 2C}
\format{
A data frame of 2,484 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (d3, p3) \cr
}
}
\source{
Sparano JA, Zhao F, Martino S, et al. Long-Term Follow-Up of the
E1199 Phase III Trial Evaluating the Role of Taxane and Schedule in
Operable Breast Cancer. J Clin Oncol 2015; 33: 2353–60.
}
\usage{
`E1199(b)_2C`
}
\description{
Kaplan-Meier digitized data from E1199(b), figure 2C (PMID 26077235). A reported sample size of 4,954 for a primary endpoint of DFS in breast cancer.
}
\examples{
summary(`E1199(b)_2C`)

kmplot(`E1199(b)_2C`)
}
\keyword{datasets}
