% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EAGLE_1}
\alias{EAGLE_1}
\title{EAGLE, figure 1}
\format{
A data frame of 368 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (bev10, bev5) \cr
}
}
\source{
Iwamoto S, Takahashi T, Tamagawa H, et al. FOLFIRI plus bevacizumab
as second-line therapy in patients with metastatic colorectal cancer
after first-line bevacizumab plus oxaliplatin-based therapy: the
randomized phase III EAGLE study. Ann Oncol 2015; 26: 1427–33.
}
\usage{
EAGLE_1
}
\description{
Kaplan-Meier digitized data from EAGLE, figure 1 (PMID 25908603). A reported sample size of 369 for a primary endpoint of PFS in colorectal cancer.
}
\examples{
summary(EAGLE_1)

kmplot(EAGLE_1)
}
\keyword{datasets}
