% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ELDA_1A}
\alias{ELDA_1A}
\title{ELDA, figure 1A}
\format{
A data frame of 299 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (cmf, wdocetaxel) \cr
}
}
\source{
Perrone F, Nuzzo F, Di Rella F, et al. Weekly docetaxel versus CMF as
adjuvant chemotherapy for older women with early breast cancer: final
results of the randomized phase III ELDA trial. Ann Oncol 2015; 26:
675–82.
}
\usage{
ELDA_1A
}
\description{
Kaplan-Meier digitized data from ELDA, figure 1A (PMID 25488686). A reported sample size of 302 for a primary endpoint of DFS in breast cancer.
}
\examples{
summary(ELDA_1A)

kmplot(ELDA_1A)
}
\keyword{datasets}
