% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ELM-PC4_2A}
\alias{ELM-PC4_2A}
\title{ELM-PC4, figure 2A}
\format{
A data frame of 1,560 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (orteronel, placebo) \cr
}
}
\source{
Saad F, Fizazi K, Jinga V, et al. Orteronel plus prednisone in
patients with chemotherapy-naive metastatic castration-resistant
prostate cancer (ELM-PC 4): a double-blind, multicentre, phase 3,
randomised, placebo-controlled trial. Lancet Oncol 2015; 16: 338–48.
}
\usage{
`ELM-PC4_2A`
}
\description{
Kaplan-Meier digitized data from ELM-PC4, figure 2A (PMID 25701170). A reported sample size of 1,560 for a primary endpoint of rPFS/OS in prostate cancer.
}
\examples{
summary(`ELM-PC4_2A`)

kmplot(`ELM-PC4_2A`)
}
\keyword{datasets}
