% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ELM-PC5_2B}
\alias{ELM-PC5_2B}
\title{ELM-PC5, figure 2B}
\format{
A data frame of 1,099 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (orteronel, placebo) \cr
}
}
\source{
Fizazi K, Jones R, Oudard S, et al. Phase III, randomized,
double-blind, multicenter trial comparing orteronel (TAK-700) plus
prednisone with placebo plus prednisone in patients with metastatic
castration-resistant prostate cancer that has progressed during or
after docetaxel-based therapy: ELM-PC 5. J Clin Oncol 2015; 33:
723–31.
}
\usage{
`ELM-PC5_2B`
}
\description{
Kaplan-Meier digitized data from ELM-PC5, figure 2B (PMID 25624429). A reported sample size of 1,099 for a primary endpoint of OS in prostate cancer.
}
\examples{
summary(`ELM-PC5_2B`)

kmplot(`ELM-PC5_2B`)
}
\keyword{datasets}
