% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ENESTg1_2B}
\alias{ENESTg1_2B}
\title{ENESTg1, figure 2B}
\format{
A data frame of 644 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (imatinib, nilotinib) \cr
}
}
\source{
Blay J-Y, Shen L, Kang Y-K, et al. Nilotinib versus imatinib as
first-line therapy for patients with unresectable or metastatic
gastrointestinal stromal tumours (ENESTg1): a randomised phase 3
trial. Lancet Oncol 2015; 16: 550–60.
}
\usage{
ENESTg1_2B
}
\description{
Kaplan-Meier digitized data from ENESTg1, figure 2B (PMID 25882987). A reported sample size of 647 for a primary endpoint of PFS in colorectal cancer.
}
\examples{
summary(ENESTg1_2B)

kmplot(ENESTg1_2B)
}
\keyword{datasets}
