% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ENSURE_3A}
\alias{ENSURE_3A}
\title{ENSURE, figure 3A}
\format{
A data frame of 217 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (erlotinib, gp) \cr
}
}
\source{
Wu Y-L, Zhou C, Liam C-K, et al. First-line erlotinib versus
gemcitabine/cisplatin in patients with advanced EGFR
mutation-positive non-small-cell lung cancer: analyses from the phase
III, randomized, open-label, ENSURE study. Ann Oncol 2015; 26:
1883–9.
}
\usage{
ENSURE_3A
}
\description{
Kaplan-Meier digitized data from ENSURE, figure 3A (PMID 26105600). A reported sample size of 217 for a primary endpoint of PFS in lung cancer.
}
\examples{
summary(ENSURE_3A)

kmplot(ENSURE_3A)
}
\keyword{datasets}
