% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EORTC22881_2}
\alias{EORTC22881_2}
\title{EORTC22881, figure 2}
\format{
A data frame of 5,318 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (boost, no boost) \cr
}
}
\source{
Bartelink H, Maingon P, Poortmans P, et al. Whole-breast irradiation
with or without a boost for patients treated with breast-conserving
surgery for early breast cancer: 20-year follow-up of a randomised
phase 3 trial. Lancet Oncol 2015; 16: 47–56.
}
\usage{
EORTC22881_2
}
\description{
Kaplan-Meier digitized data from EORTC22881, figure 2 (PMID 25500422). A reported sample size of 2,657 for a primary endpoint of OS in breast cancer.
}
\examples{
summary(EORTC22881_2)

kmplot(EORTC22881_2)
}
\keyword{datasets}
