% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ERCC1_2B}
\alias{ERCC1_2B}
\title{ERCC1, figure 2B}
\format{
A data frame of 468 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (cisplatin_pemetrexed, p1) \cr
}
}
\source{
Lee SM, Falzon M, Blackhall F, et al. Randomized Prospective
Biomarker Trial of ERCC1 for Comparing Platinum and Nonplatinum
Therapy in Advanced Non-Small-Cell Lung Cancer: ERCC1 Trial (ET). J
Clin Oncol 2017; 35: 402–11.
}
\usage{
ERCC1_2B
}
\description{
Kaplan-Meier digitized data from ERCC1, figure 2B (PMID 27893326). A reported sample size of 648 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(ERCC1_2B)

kmplot(ERCC1_2B)
}
\keyword{datasets}
