% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FFCD2001_S1A}
\alias{FFCD2001_S1A}
\title{FFCD2001, figure S1A}
\format{
A data frame of 282 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (fu, iri) \cr
}
}
\source{
Aparicio T, Lavau-Denes S, Phelip JM, et al. Randomized phase III
trial in elderly patients comparing LV5FU2 with or without irinotecan
for first-line treatment of metastatic colorectal cancer (FFCD
2001-02). Ann Oncol 2016; 27: 121–7.
}
\usage{
FFCD2001_S1A
}
\description{
Kaplan-Meier digitized data from FFCD2001, figure S1A (PMID 26487578). A reported sample size of 212 for a primary endpoint of PFS in colorectal cancer.
}
\examples{
summary(FFCD2001_S1A)

kmplot(FFCD2001_S1A)
}
\keyword{datasets}
