% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FIRE3_2B}
\alias{FIRE3_2B}
\title{FIRE3, figure 2B}
\format{
A data frame of 592 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (folfiri_bevacizumab, folfiri_cetuximab) \cr
}
}
\source{
Heinemann V, von Weikersthal LF, Decker T, et al. FOLFIRI plus
cetuximab versus FOLFIRI plus bevacizumab as first-line treatment for
patients with metastatic colorectal cancer (FIRE-3): a randomised,
open-label, phase 3 trial. Lancet Oncol 2014; 15: 1065–75.
}
\usage{
FIRE3_2B
}
\description{
Kaplan-Meier digitized data from FIRE3, figure 2B (PMID 25088940). A reported sample size of 592 for a primary endpoint of objective_response in colorectal cancer.
}
\examples{
summary(FIRE3_2B)

kmplot(FIRE3_2B)
}
\keyword{datasets}
