% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GEICAM2003_2B}
\alias{GEICAM2003_2B}
\title{GEICAM2003, figure 2B}
\format{
A data frame of 1,384 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (ec_t, et_x) \cr
}
}
\source{
Martín M, Ruiz Simón A, Ruiz Borrego M, et al. Epirubicin Plus
Cyclophosphamide Followed by Docetaxel Versus Epirubicin Plus
Docetaxel Followed by Capecitabine As Adjuvant Therapy for
Node-Positive Early Breast Cancer: Results From the GEICAM/2003-10
Study. J Clin Oncol 2015; 33: 3788–95.
}
\usage{
GEICAM2003_2B
}
\description{
Kaplan-Meier digitized data from GEICAM2003, figure 2B (PMID 26416999). A reported sample size of 1,384 for a primary endpoint of iDFS in breast cancer.
}
\examples{
summary(GEICAM2003_2B)

kmplot(GEICAM2003_2B)
}
\keyword{datasets}
