% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GETUG12_2}
\alias{GETUG12_2}
\title{GETUG12, figure 2}
\format{
A data frame of 413 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab RFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (adt, adt_de) \cr
}
}
\source{
Fizazi K, Faivre L, Lesaunier F, et al. Androgen deprivation therapy
plus docetaxel and estramustine versus androgen deprivation therapy
alone for high-risk localised prostate cancer (GETUG 12): a phase 3
randomised controlled trial. Lancet Oncol 2015; 16: 787–94.
}
\usage{
GETUG12_2
}
\description{
Kaplan-Meier digitized data from GETUG12, figure 2 (PMID 26028518). A reported sample size of 207 for a primary endpoint of RFS in prostate cancer.
}
\examples{
summary(GETUG12_2)

kmplot(GETUG12_2)
}
\keyword{datasets}
