% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GETUG16_2}
\alias{GETUG16_2}
\title{GETUG16, figure 2}
\format{
A data frame of 730 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (rt, rt_goserelin) \cr
}
}
\source{
Carrie C, Hasbini A, de Laroche G, et al. Salvage radiotherapy with
or without short-term hormone therapy for rising prostate-specific
antigen concentration after radical prostatectomy (GETUG-AFU 16): a
randomised, multicentre, open-label phase 3 trial. Lancet Oncol 2016;
17: 747–56.
}
\usage{
GETUG16_2
}
\description{
Kaplan-Meier digitized data from GETUG16, figure 2 (PMID 27160475). A reported sample size of 743 for a primary endpoint of PFS in prostate cancer.
}
\examples{
summary(GETUG16_2)

kmplot(GETUG16_2)
}
\keyword{datasets}
