% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GIM2_2C}
\alias{GIM2_2C}
\title{GIM2, figure 2C}
\format{
A data frame of 2,003 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (q2, q3) \cr
}
}
\source{
Del Mastro L, De Placido S, Bruzzi P, et al. Fluorouracil and
dose-dense chemotherapy in adjuvant treatment of patients with
early-stage breast cancer: an open-label, 2 × 2 factorial, randomised
phase 3 trial. Lancet 2015; 385: 1863–72.
}
\usage{
GIM2_2C
}
\description{
Kaplan-Meier digitized data from GIM2, figure 2C (PMID 25740286). A reported sample size of 2,091 for a primary endpoint of DFS in breast cancer.
}
\examples{
summary(GIM2_2C)

kmplot(GIM2_2C)
}
\keyword{datasets}
