% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GIM_3}
\alias{GIM_3}
\title{GIM, figure 3}
\format{
A data frame of 281 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (control, lhrha) \cr
}
}
\source{
Lambertini M, Boni L, Michelotti A, et al. Ovarian Suppression With
Triptorelin During Adjuvant Breast Cancer Chemotherapy and Long-term
Ovarian Function, Pregnancies, and Disease-Free Survival: A
Randomized Clinical Trial. JAMA 2015; 314: 2632–40.
}
\usage{
GIM_3
}
\description{
Kaplan-Meier digitized data from GIM, figure 3 (PMID 26720025). A reported sample size of 281 for a primary endpoint of Incidence_chemo-induced menopause in breast cancer.
}
\examples{
summary(GIM_3)

kmplot(GIM_3)
}
\keyword{datasets}
