% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HYPRO_2B}
\alias{HYPRO_2B}
\title{HYPRO, figure 2B}
\format{
A data frame of 804 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (conventional fractionated, hypofractionated) \cr
}
}
\source{
Incrocci L, Wortel RC, Alemayehu WG, et al. Hypofractionated versus
conventionally fractionated radiotherapy for patients with localised
prostate cancer (HYPRO): final efficacy results from a randomised,
multicentre, open-label, phase 3 trial. Lancet Oncol 2016; 17:
1061–9.
}
\usage{
HYPRO_2B
}
\description{
Kaplan-Meier digitized data from HYPRO, figure 2B (PMID 27339116). A reported sample size of 804 for a primary endpoint of RFS in prostate cancer.
}
\examples{
summary(HYPRO_2B)

kmplot(HYPRO_2B)
}
\keyword{datasets}
