% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IBCSG2200_2A}
\alias{IBCSG2200_2A}
\title{IBCSG2200, figure 2A}
\format{
A data frame of 1,081 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (cm, no cm) \cr
}
}
\source{
Colleoni M, Gray KP, Gelber S, et al. Low-Dose Oral Cyclophosphamide
and Methotrexate Maintenance for Hormone Receptor-Negative Early
Breast Cancer: International Breast Cancer Study Group Trial 22-00. J
Clin Oncol 2016; 34: 3400–8.
}
\usage{
IBCSG2200_2A
}
\description{
Kaplan-Meier digitized data from IBCSG2200, figure 2A (PMID 27325862). A reported sample size of 1,086 for a primary endpoint of DFS in breast cancer.
}
\examples{
summary(IBCSG2200_2A)

kmplot(IBCSG2200_2A)
}
\keyword{datasets}
