% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMELDA_2B}
\alias{IMELDA_2B}
\title{IMELDA, figure 2B}
\format{
A data frame of 185 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (bevacizumab, bevacizumab_capecitabine) \cr
}
}
\source{
Gligorov J, Doval D, Bines J, et al. Maintenance capecitabine and
bevacizumab versus bevacizumab alone after initial first-line
bevacizumab and docetaxel for patients with HER2-negative metastatic
breast cancer (IMELDA): a randomised, open-label, phase 3 trial.
Lancet Oncol 2014; 15: 1351–60.
}
\usage{
IMELDA_2B
}
\description{
Kaplan-Meier digitized data from IMELDA, figure 2B (PMID 25273343). A reported sample size of 284 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(IMELDA_2B)

kmplot(IMELDA_2B)
}
\keyword{datasets}
