% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{INSPIRE_2B}
\alias{INSPIRE_2B}
\title{INSPIRE, figure 2B}
\format{
A data frame of 633 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (ct, necitumumab_ct) \cr
}
}
\source{
Paz-Ares L, Mezger J, Ciuleanu TE, et al. Necitumumab plus pemetrexed
and cisplatin as first-line therapy in patients with stage IV
non-squamous non-small-cell lung cancer (INSPIRE): an open-label,
randomised, controlled phase 3 study. Lancet Oncol 2015; 16: 328–37.
}
\usage{
INSPIRE_2B
}
\description{
Kaplan-Meier digitized data from INSPIRE, figure 2B (PMID 25701171). A reported sample size of 633 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(INSPIRE_2B)

kmplot(INSPIRE_2B)
}
\keyword{datasets}
