% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{INT0142_2B}
\alias{INT0142_2B}
\title{INT0142, figure 2B}
\format{
A data frame of 337 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (tamoxifen, tamoxifen_ofs) \cr
}
}
\source{
Tevaarwerk AJ, Wang M, Zhao F, et al. Phase III comparison of
tamoxifen versus tamoxifen plus ovarian function suppression in
premenopausal women with node-negative, hormone receptor-positive
breast cancer (E-3193, INT-0142): a trial of the Eastern Cooperative
Oncology Group. J Clin Oncol 2014; 32: 3948–58.
}
\usage{
INT0142_2B
}
\description{
Kaplan-Meier digitized data from INT0142, figure 2B (PMID 25349302). A reported sample size of 345 for a primary endpoint of DFS/OS in breast cancer.
}
\examples{
summary(INT0142_2B)

kmplot(INT0142_2B)
}
\keyword{datasets}
