% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipd.R
\name{ipd}
\alias{ipd}
\title{IPD}
\usage{
ipd(
  time,
  prob,
  t.atrisk,
  n.atrisk,
  total = NA,
  arm = "arm",
  tau = max(t.atrisk),
  interpolate = FALSE
)
}
\arguments{
\item{time, prob}{vectors of survival time and probabilities}

\item{t.atrisk, n.atrisk}{vectors of at-risk times and number at-risk}

\item{total}{(optional) total number of events}

\item{arm}{(optional) arm label}

\item{tau}{maximum follow-up time}

\item{interpolate}{logical or integer; if \code{TRUE} or an integer,
constant interpolation will be performed for the time-probability
vectors; this may be helpful in situations with a small number of
digitized points or if the censoring distribution is difficult to
estimate}
}
\description{
Generate individual patient data (IPD) from digitized survival curve data
using the method described by Guyot (2012).
}
\examples{
\dontrun{
xy <- system.file(
  'etc', 'init', 'Checkmate_067_S3A_Nivolumab.csv',
  package = 'kmdata'
)
ar <- system.file(
  'etc', 'init', 'Checkmate_067_At_Risk.csv',
  package = 'kmdata'
)

dd <- read.csv(xy)
aa <- read.csv(ar)
aa <- aa[aa$nrisk > 0, ]

## use full at-risk table data
ipd1 <- ipd(dd$T, dd$S, aa$trisk, aa$nrisk, arm = 'Nivo')
kmplot(ipd1, xaxis.at = 0:4 * 10, xlim = c(0, 45))


## only using 1/2 of at-risk table
idx <- replace(seq_along(aa$trisk) \%\% 2 == 1, nrow(aa), TRUE)
ipd2 <- ipd(dd$T, dd$S, aa$trisk[idx], aa$nrisk[idx], arm = 'Nivo')
kmplot(ipd2, xaxis.at = 0:4 * 10, xlim = c(0, 45))


## only using 1/4 of at-risk table
idx <- replace(seq_along(aa$trisk) \%\% 4 == 1, nrow(aa), TRUE)
ipd3 <- ipd(dd$T, dd$S, aa$trisk[idx], aa$nrisk[idx], arm = 'Nivo')
kmplot(ipd3, xaxis.at = 0:4 * 10, xlim = c(0, 45))


## only using 1/8 of at-risk table
idx <- replace(seq_along(aa$trisk) \%\% 8 == 1, nrow(aa), TRUE)
ipd4 <- ipd(dd$T, dd$S, aa$trisk[idx], aa$nrisk[idx], arm = 'Nivo')
kmplot(ipd4, xaxis.at = 0:4 * 10, xlim = c(0, 45))
}

}
\references{
Guyot, Patricia, et al. Enhanced secondary analysis of survival data:
reconstructing the data from published Kaplan-Meier survival curves.
\emph{BMC Medical Research Methodology} 2012, \strong{12}:9.
}
