% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ITACa_2A}
\alias{ITACa_2A}
\title{ITACa, figure 2A}
\format{
A data frame of 370 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (ct, ct_b) \cr
}
}
\source{
Passardi A, Nanni O, Tassinari D, et al. Effectiveness of bevacizumab
added to standard chemotherapy in metastatic colorectal cancer: final
results for first-line treatment from the ITACa randomized clinical
trial. Ann Oncol 2015; 26: 1201–7.
}
\usage{
ITACa_2A
}
\description{
Kaplan-Meier digitized data from ITACa, figure 2A (PMID 25735317). A reported sample size of 376 for a primary endpoint of PFS in colorectal cancer.
}
\examples{
summary(ITACa_2A)

kmplot(ITACa_2A)
}
\keyword{datasets}
