% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{JCOG0509_2A}
\alias{JCOG0509_2A}
\title{JCOG0509, figure 2A}
\format{
A data frame of 284 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (amrubicin_cisplatin, irinotecan_cisplatin) \cr
}
}
\source{
Satouchi M, Kotani Y, Shibata T, et al. Phase III study comparing
amrubicin plus cisplatin with irinotecan plus cisplatin in the
treatment of extensive-disease small-cell lung cancer: JCOG 0509. J
Clin Oncol 2014; 32: 1262–8.
}
\usage{
JCOG0509_2A
}
\description{
Kaplan-Meier digitized data from JCOG0509, figure 2A (PMID 24638015). A reported sample size of 284 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(JCOG0509_2A)

kmplot(JCOG0509_2A)
}
\keyword{datasets}
