% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{JCOG0605_2A}
\alias{JCOG0605_2A}
\title{JCOG0605, figure 2A}
\format{
A data frame of 180 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (combination_chemotherapy, topotecan) \cr
}
}
\source{
Goto K, Ohe Y, Shibata T, et al. Combined chemotherapy with
cisplatin, etoposide, and irinotecan versus topotecan alone as
second-line treatment for patients with sensitive relapsed small-cell
lung cancer (JCOG0605): a multicentre, open-label, randomised phase 3
trial. Lancet Oncol 2016; 17: 1147–57.
}
\usage{
JCOG0605_2A
}
\description{
Kaplan-Meier digitized data from JCOG0605, figure 2A (PMID 27312053). A reported sample size of 180 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(JCOG0605_2A)

kmplot(JCOG0605_2A)
}
\keyword{datasets}
