% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KEYNOTE024_2}
\alias{KEYNOTE024_2}
\title{KEYNOTE024, figure 2}
\format{
A data frame of 305 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (chemo, pembrolizumab) \cr
}
}
\source{
Reck M, Rodríguez-Abreu D, Robinson AG, et al. Pembrolizumab versus
Chemotherapy for PD-L1-Positive Non-Small-Cell Lung Cancer. N Engl J
Med 2016; 375: 1823–33.
}
\usage{
KEYNOTE024_2
}
\description{
Kaplan-Meier digitized data from KEYNOTE024, figure 2 (PMID 27718847). A reported sample size of 305 for a primary endpoint of PFS in lung cancer.
}
\examples{
summary(KEYNOTE024_2)

kmplot(KEYNOTE024_2)
}
\keyword{datasets}
