% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-demo.R
\docType{data}
\name{kmdata_demo}
\alias{kmdata_demo}
\title{\code{kmdata} demographics data}
\format{
A data frame of 331 observations and 12 variables:

\tabular{lll}{

\tab \code{Publication} \tab publication identifier \cr
\tab \code{Therapy} \tab therapy received \cr
\tab \code{Description} \tab additional information about \code{Therapy}
  where appropriate \cr
\tab \code{Median Age} \tab the median age by publication/therapy \cr
\tab \code{Age Range} \tab the age range by publication/therapy \cr
\tab \code{Sex:Males} \tab the number of males by publication/therapy \cr
\tab \code{Sex:Females} \tab the number of females by publication/therapy \cr
\tab \code{Race:White} \tab the number of White/Caucasian by
  publication/therapy \cr
\tab \code{Race:Other} \tab the number of non-White/Caucasian by
  publication/therapy \cr
\tab \code{ECOG:0/1} \tab the number of patients with an ECOG performance
  score of 0 or 1 by publication/therapy \cr
\tab \code{ECOG:2+} \tab the number of patients with an ECOG performance
  score of 2+ by publication/therapy \cr
\tab \code{ECOG:UNK} \tab the number of patients with an ECOG performance
  score of unknown by publication/therapy \cr

}
}
\usage{
kmdata_demo
}
\description{
Demographics of data sets in \code{kmdata} package by treatment arm. Missing
values are shown when either the publication did not report or the number
could not be inferred from available information.
}
\seealso{
\code{\link{kmdata_key}}; \code{\link{select_kmdata}};
\code{\link{summary.kmdata}}
}
\keyword{datasets}
