% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmm.R
\name{kmm}
\alias{kmm}
\title{Kaplan-Meier metrics}
\usage{
kmm(x, y, times = NULL)
}
\arguments{
\item{x, y}{\code{survfit} objects}

\item{times}{times to evaluate survival}
}
\description{
Compute metrics between two \code{\link[survival]{survfit}} objects; see
examples in \code{\link{kmm_legend}}.
}
