% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{kmplot}
\alias{kmplot}
\alias{plot.kmdata}
\title{kmplot}
\usage{
kmplot(
  x,
  ...,
  relevel = FALSE,
  col = NULL,
  plot = TRUE,
  xlim = NULL,
  ylim = NULL,
  xaxis.at = NULL,
  xlab = NULL,
  ylab = NULL,
  lr_test = TRUE,
  test_details = TRUE,
  median = TRUE,
  atrisk = TRUE,
  mark.time = TRUE,
  title = NULL,
  legend = TRUE
)

\method{plot}{kmdata}(
  x,
  ...,
  relevel = FALSE,
  col = NULL,
  plot = TRUE,
  xlim = NULL,
  ylim = NULL,
  xaxis.at = NULL,
  xlab = NULL,
  ylab = NULL,
  lr_test = TRUE,
  test_details = TRUE,
  median = TRUE,
  atrisk = TRUE,
  mark.time = TRUE,
  title = NULL,
  legend = TRUE
)
}
\arguments{
\item{x}{a data set of class \code{"kmdata"}, i.e., one of the data sets
in the \pkg{kmdata} package; alternatively, a data frame with columns
"time", "event", and "arm"}

\item{...}{additional arguments passed to \code{\link[survival]{plot.survfit}}
or further to \code{\link{plot.default}} or \code{\link{par}}}

\item{relevel}{logical; if \code{TRUE}, group order is reversed; by default,
the arms are in alphabetical order which may not be desired for some
placebo or control arms}

\item{col}{a vector of colors for the survival curves (recycled for at-risk
table, medians, and legend)}

\item{plot}{logical; if \code{TRUE}, a KM figure is drawn}

\item{xlim, ylim}{x- and y-axis limits}

\item{xaxis.at}{x-axis positions of ticks and at-risk table}

\item{xlab, ylab}{the x- and y-axis labels}

\item{lr_test}{logical; if \code{TRUE}, log-rank test is shown in upper
right corner of figure}

\item{test_details}{logical; if \code{TRUE}, test statistic and degrees
of freedom are added with p-value for log-rank test}

\item{median}{logical; if \code{TRUE}, the medians for each curve is added}

\item{atrisk}{logical; if \code{TRUE}, an at-risk table is drawn below plot}

\item{mark.time}{passed to \code{\link[survival]{plot.survfit}}}

\item{title}{optional title for plot; default is \code{attr(data, 'title')}}

\item{legend}{logical; if \code{TRUE}, a legend for the arms, total events,
and hazard ratios is added}
}
\description{
Make a Kaplan-Meier plot of a data set.
}
\examples{
kmplot(ATTENTION_2A)

kmplot(
  ATTENTION_2A,
  relevel = TRUE, median = FALSE,
  col = 3:4, xaxis.at = 0:4 * 6
)

## or equivalently
s <- survfit(Surv(time, event) ~ arm, ATTENTION_2A)
plot(s, col = 1:2, mark.time = TRUE)


## kmplot can be used generically given the proper data structure
dat <- data.frame(time = aml$time, event = aml$status, arm = aml$x)
kmplot(dat)

}
