% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LUMELung1_2A}
\alias{LUMELung1_2A}
\title{LUMELung1, figure 2A}
\format{
A data frame of 1,134 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (docetaxel_nintedanib, docetaxel_placebo) \cr
}
}
\source{
Reck M, Kaiser R, Mellemgaard A, et al. Docetaxel plus nintedanib
versus docetaxel plus placebo in patients with previously treated
non-small-cell lung cancer (LUME-Lung 1): a phase 3, double-blind,
randomised controlled trial. Lancet Oncol 2014; 15: 143–55.
}
\usage{
LUMELung1_2A
}
\description{
Kaplan-Meier digitized data from LUMELung1, figure 2A (PMID 24411639). A reported sample size of 655 for a primary endpoint of PFS in lung cancer.
}
\examples{
summary(LUMELung1_2A)

kmplot(LUMELung1_2A)
}
\keyword{datasets}
