% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LUXBreast1_4}
\alias{LUXBreast1_4}
\title{LUXBreast1, figure 4}
\format{
A data frame of 508 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (afatinib, t) \cr
}
}
\source{
Harbeck N, Huang C-S, Hurvitz S, et al. Afatinib plus vinorelbine
versus trastuzumab plus vinorelbine in patients with
HER2-overexpressing metastatic breast cancer who had progressed on
one previous trastuzumab treatment (LUX-Breast 1): an open-label,
randomised, phase 3 trial. Lancet Oncol 2016; 17: 357–66.
}
\usage{
LUXBreast1_4
}
\description{
Kaplan-Meier digitized data from LUXBreast1, figure 4 (PMID 26822398). A reported sample size of 508 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(LUXBreast1_4)

kmplot(LUXBreast1_4)
}
\keyword{datasets}
