% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LUXLung3-6_2B}
\alias{LUXLung3-6_2B}
\title{LUXLung3-6, figure 2B}
\format{
A data frame of 364 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (afatinib, gemicitabine-cisplatin) \cr
}
}
\source{
Yang JC-H, Wu Y-L, Schuler M, et al. Afatinib versus cisplatin-based
chemotherapy for EGFR mutation-positive lung adenocarcinoma (LUX-Lung
3 and LUX-Lung 6): analysis of overall survival data from two
randomised, phase 3 trials. Lancet Oncol 2015; 16: 141–51.
}
\usage{
`LUXLung3-6_2B`
}
\description{
Kaplan-Meier digitized data from LUXLung3-6, figure 2B (PMID 25589191). A reported sample size of 345,364 for a primary endpoint of PFS in lung cancer.
}
\examples{
summary(`LUXLung3-6_2B`)

kmplot(`LUXLung3-6_2B`)
}
\keyword{datasets}
