% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LUXLung5_1B}
\alias{LUXLung5_1B}
\title{LUXLung5, figure 1B}
\format{
A data frame of 202 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (afatinib_paclitaxel, chemo) \cr
}
}
\source{
Schuler M, Yang JC-H, Park K, et al. Afatinib beyond progression in
patients with non-small-cell lung cancer following chemotherapy,
erlotinib/gefitinib and afatinib: phase III randomized LUX-Lung 5
trial. Ann Oncol 2016; 27: 417–23.
}
\usage{
LUXLung5_1B
}
\description{
Kaplan-Meier digitized data from LUXLung5, figure 1B (PMID 26646759). A reported sample size of 220 for a primary endpoint of PFS in lung cancer.
}
\examples{
summary(LUXLung5_1B)

kmplot(LUXLung5_1B)
}
\keyword{datasets}
