% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LUXLung6(2)_2A}
\alias{LUXLung6(2)_2A}
\title{LUXLung6(2), figure 2A}
\format{
A data frame of 364 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (afatinib, gemcitabine_cisplatin) \cr
}
}
\source{
Wu Y-L, Zhou C, Hu C-P, et al. Afatinib versus cisplatin plus
gemcitabine for first-line treatment of Asian patients with advanced
non-small-cell lung cancer harbouring EGFR mutations (LUX-Lung 6): an
open-label, randomised phase 3 trial. Lancet Oncol 2014; 15: 213–22.
}
\usage{
`LUXLung6(2)_2A`
}
\description{
Kaplan-Meier digitized data from LUXLung6(2), figure 2A (PMID 24439929). A reported sample size of 364 for a primary endpoint of PFS in lung cancer.
}
\examples{
summary(`LUXLung6(2)_2A`)

kmplot(`LUXLung6(2)_2A`)
}
\keyword{datasets}
