% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LUXLung8_3A}
\alias{LUXLung8_3A}
\title{LUXLung8, figure 3A}
\format{
A data frame of 795 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (afatinib, erlotinib) \cr
}
}
\source{
Soria J-C, Felip E, Cobo M, et al. Afatinib versus erlotinib as
second-line treatment of patients with advanced squamous cell
carcinoma of the lung (LUX-Lung 8): an open-label randomised
controlled phase 3 trial. Lancet Oncol 2015; 16: 897–907.
}
\usage{
LUXLung8_3A
}
\description{
Kaplan-Meier digitized data from LUXLung8, figure 3A (PMID 26156651). A reported sample size of 795 for a primary endpoint of PFS in lung cancer.
}
\examples{
summary(LUXLung8_3A)

kmplot(LUXLung8_3A)
}
\keyword{datasets}
