% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MA17R_1B}
\alias{MA17R_1B}
\title{MA17R, figure 1B}
\format{
A data frame of 1,918 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (l, placebo) \cr
}
}
\source{
Goss PE, Ingle JN, Pritchard KI, et al. Extending Aromatase-Inhibitor
Adjuvant Therapy to 10 Years. N Engl J Med 2016; 375: 209–19.
}
\usage{
MA17R_1B
}
\description{
Kaplan-Meier digitized data from MA17R, figure 1B (PMID 27264120). A reported sample size of 1,918 for a primary endpoint of DFS in breast cancer.
}
\examples{
summary(MA17R_1B)

kmplot(MA17R_1B)
}
\keyword{datasets}
