% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MA31_2}
\alias{MA31_2}
\title{MA31, figure 2}
\format{
A data frame of 652 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (ltax_l, ttax_t) \cr
}
}
\source{
Gelmon KA, Boyle FM, Kaufman B, et al. Lapatinib or Trastuzumab Plus
Taxane Therapy for Human Epidermal Growth Factor Receptor 2-Positive
Advanced Breast Cancer: Final Results of NCIC CTG MA.31. J Clin Oncol
2015; 33: 1574–83.
}
\usage{
MA31_2
}
\description{
Kaplan-Meier digitized data from MA31, figure 2 (PMID 25779558). A reported sample size of 652 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(MA31_2)

kmplot(MA31_2)
}
\keyword{datasets}
