% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAGRIT_2A}
\alias{MAGRIT_2A}
\title{MAGRIT, figure 2A}
\format{
A data frame of 2,272 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (mage_a3, placebo) \cr
}
}
\source{
Vansteenkiste JF, Cho BC, Vanakesa T, et al. Efficacy of the MAGE-A3
cancer immunotherapeutic as adjuvant therapy in patients with
resected MAGE-A3-positive non-small-cell lung cancer (MAGRIT): a
randomised, double-blind, placebo-controlled, phase 3 trial. Lancet
Oncol 2016; 17: 822–35.
}
\usage{
MAGRIT_2A
}
\description{
Kaplan-Meier digitized data from MAGRIT, figure 2A (PMID 27132212). A reported sample size of 2,312 for a primary endpoint of DFS(3 primary in diff pop) in lung cancer.
}
\examples{
summary(MAGRIT_2A)

kmplot(MAGRIT_2A)
}
\keyword{datasets}
