% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAINSAIL_2B}
\alias{MAINSAIL_2B}
\title{MAINSAIL, figure 2B}
\format{
A data frame of 1,059 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (lenolidomide, placebo) \cr
}
}
\source{
Petrylak DP, Vogelzang NJ, Budnik N, et al. Docetaxel and prednisone
with or without lenalidomide in chemotherapy-naive patients with
metastatic castration-resistant prostate cancer (MAINSAIL): a
randomised, double-blind, placebo-controlled phase 3 trial. Lancet
Oncol 2015; 16: 417–25.
}
\usage{
MAINSAIL_2B
}
\description{
Kaplan-Meier digitized data from MAINSAIL, figure 2B (PMID 25743937). A reported sample size of 1,059 for a primary endpoint of OS in prostate cancer.
}
\examples{
summary(MAINSAIL_2B)

kmplot(MAINSAIL_2B)
}
\keyword{datasets}
