% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAPS_2C}
\alias{MAPS_2C}
\title{MAPS, figure 2C}
\format{
A data frame of 448 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (pc, pcb) \cr
}
}
\source{
Zalcman G, Mazieres J, Margery J, et al. Bevacizumab for newly
diagnosed pleural mesothelioma in the Mesothelioma Avastin Cisplatin
Pemetrexed Study (MAPS): a randomised, controlled, open-label, phase
3 trial. Lancet 2016; 387: 1405–14.
}
\usage{
MAPS_2C
}
\description{
Kaplan-Meier digitized data from MAPS, figure 2C (PMID 26719230). A reported sample size of 448 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(MAPS_2C)

kmplot(MAPS_2C)
}
\keyword{datasets}
