% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MARIANNE(a)_2A}
\alias{MARIANNE(a)_2A}
\title{MARIANNE(a), figure 2A}
\format{
A data frame of 732 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (t-dm1, trastuzumab_taxane) \cr
}
}
\source{
Perez EA, Barrios C, Eiermann W, et al. Trastuzumab Emtansine With or
Without Pertuzumab Versus Trastuzumab Plus Taxane for Human Epidermal
Growth Factor Receptor 2-Positive, Advanced Breast Cancer: Primary
Results From the Phase III MARIANNE Study. J Clin Oncol 2017; 35:
141–8.
}
\usage{
`MARIANNE(a)_2A`
}
\description{
Kaplan-Meier digitized data from MARIANNE(a), figure 2A (PMID 28056202). A reported sample size of 1,095 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(`MARIANNE(a)_2A`)

kmplot(`MARIANNE(a)_2A`)
}
\keyword{datasets}
