% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MARQUEE_2B}
\alias{MARQUEE_2B}
\title{MARQUEE, figure 2B}
\format{
A data frame of 1,048 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (erlotinib_placebo, erlotinib_tivantinib) \cr
}
}
\source{
Scagliotti G, von Pawel J, Novello S, et al. Phase III Multinational,
Randomized, Double-Blind, Placebo-Controlled Study of Tivantinib (ARQ
197) Plus Erlotinib Versus Erlotinib Alone in Previously Treated
Patients With Locally Advanced or Metastatic Nonsquamous
Non-Small-Cell Lung Cancer. J Clin Oncol 2015; 33: 2667–74.
}
\usage{
MARQUEE_2B
}
\description{
Kaplan-Meier digitized data from MARQUEE, figure 2B (PMID 26169611). A reported sample size of 1,048 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(MARQUEE_2B)

kmplot(MARQUEE_2B)
}
\keyword{datasets}
