% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METlung_3A}
\alias{METlung_3A}
\title{METlung, figure 3A}
\format{
A data frame of 499 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (onaturzumab_erlotinib, placebo_erlotinib) \cr
}
}
\source{
Spigel DR, Edelman MJ, O’Byrne K, et al. Results From the Phase III
Randomized Trial of Onartuzumab Plus Erlotinib Versus Erlotinib in
Previously Treated Stage IIIB or IV Non-Small-Cell Lung Cancer:
METLung. J Clin Oncol 2017; 35: 412–20.
}
\usage{
METlung_3A
}
\description{
Kaplan-Meier digitized data from METlung, figure 3A (PMID 27937096). A reported sample size of 499 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(METlung_3A)

kmplot(METlung_3A)
}
\keyword{datasets}
