% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MINDACT_2E}
\alias{MINDACT_2E}
\title{MINDACT, figure 2E}
\format{
A data frame of 690 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (chemo, no_chemotherapy) \cr
}
}
\source{
Cardoso F, van’t Veer LJ, Bogaerts J, et al. 70-Gene Signature as an
Aid to Treatment Decisions in Early-Stage Breast Cancer. N Engl J Med
2016; 375: 717–29.
}
\usage{
MINDACT_2E
}
\description{
Kaplan-Meier digitized data from MINDACT, figure 2E (PMID 27557300). A reported sample size of 1,550 for a primary endpoint of DMFS in breast cancer.
}
\examples{
summary(MINDACT_2E)

kmplot(MINDACT_2E)
}
\keyword{datasets}
