% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MIROX_2B}
\alias{MIROX_2B}
\title{MIROX, figure 2B}
\format{
A data frame of 282 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (folfox4, folfox7_folfiri) \cr
}
}
\source{
Hebbar M, Chibaudel B, André T, et al. FOLFOX4 versus sequential
dose-dense FOLFOX7 followed by FOLFIRI in patients with resectable
metastatic colorectal cancer (MIROX): a pragmatic approach to
chemotherapy timing with perioperative or postoperative chemotherapy
from an open-label, randomized phase III trial. Ann Oncol 2015; 26:
340–7.
}
\usage{
MIROX_2B
}
\description{
Kaplan-Meier digitized data from MIROX, figure 2B (PMID 25403578). A reported sample size of 284 for a primary endpoint of DFS-2yr in colorectal cancer.
}
\examples{
summary(MIROX_2B)

kmplot(MIROX_2B)
}
\keyword{datasets}
