% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MONALEESA2_1}
\alias{MONALEESA2_1}
\title{MONALEESA2, figure 1}
\format{
A data frame of 668 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (placebo, ribociclib_letrozole) \cr
}
}
\source{
Hortobagyi GN, Stemmer SM, Burris HA, et al. Ribociclib as First-Line
Therapy for HR-Positive, Advanced Breast Cancer. N Engl J Med 2016;
375: 1738–48.
}
\usage{
MONALEESA2_1
}
\description{
Kaplan-Meier digitized data from MONALEESA2, figure 1 (PMID 27717303). A reported sample size of 668 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(MONALEESA2_1)

kmplot(MONALEESA2_1)
}
\keyword{datasets}
