% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MRC-RT01_2C}
\alias{MRC-RT01_2C}
\title{MRC-RT01, figure 2C}
\format{
A data frame of 835 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (control, escalated) \cr
}
}
\source{
Dearnaley DP, Jovic G, Syndikus I, et al. Escalated-dose versus
control-dose conformal radiotherapy for prostate cancer: long-term
results from the MRC RT01 randomised controlled trial. Lancet Oncol
2014; 15: 464–73.
}
\usage{
`MRC-RT01_2C`
}
\description{
Kaplan-Meier digitized data from MRC-RT01, figure 2C (PMID 24581940). A reported sample size of 862 for a primary endpoint of bPFS/OS in prostate cancer.
}
\examples{
summary(`MRC-RT01_2C`)

kmplot(`MRC-RT01_2C`)
}
\keyword{datasets}
