% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCICBR26_2B}
\alias{NCICBR26_2B}
\title{NCICBR26, figure 2B}
\format{
A data frame of 720 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (dacomitinib, placebo) \cr
}
}
\source{
Ellis PM, Shepherd FA, Millward M, et al. Dacomitinib compared with
placebo in pretreated patients with advanced or metastatic
non-small-cell lung cancer (NCIC CTG BR.26): a double-blind,
randomised, phase 3 trial. Lancet Oncol 2014; 15: 1379–88.
}
\usage{
NCICBR26_2B
}
\description{
Kaplan-Meier digitized data from NCICBR26, figure 2B (PMID 25439692). A reported sample size of 720 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(NCICBR26_2B)

kmplot(NCICBR26_2B)
}
\keyword{datasets}
