% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCT00294996_1B}
\alias{NCT00294996_1B}
\title{NCT00294996, figure 1B}
\format{
A data frame of 363 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (mpt, pt) \cr
}
}
\source{
Baselga J, Manikhas A, Cortés J, et al. Phase III trial of
nonpegylated liposomal doxorubicin in combination with trastuzumab
and paclitaxel in HER2-positive metastatic breast cancer. Ann Oncol
2014; 25: 592–8.
}
\usage{
NCT00294996_1B
}
\description{
Kaplan-Meier digitized data from NCT00294996, figure 1B (PMID 24401928). A reported sample size of 181 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(NCT00294996_1B)

kmplot(NCT00294996_1B)
}
\keyword{datasets}
