% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCT00337103_2B}
\alias{NCT00337103_2B}
\title{NCT00337103, figure 2B}
\format{
A data frame of 1,102 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (cap, eribulin) \cr
}
}
\source{
Kaufman PA, Awada A, Twelves C, et al. Phase III open-label
randomized study of eribulin mesylate versus capecitabine in patients
with locally advanced or metastatic breast cancer previously treated
with an anthracycline and a taxane. J Clin Oncol 2015; 33: 594–601.
}
\usage{
NCT00337103_2B
}
\description{
Kaplan-Meier digitized data from NCT00337103, figure 2B (PMID 25605862). A reported sample size of 1,102 for a primary endpoint of OS/PFS in breast cancer.
}
\examples{
summary(NCT00337103_2B)

kmplot(NCT00337103_2B)
}
\keyword{datasets}
