% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCT00596830_2A}
\alias{NCT00596830_2A}
\title{NCT00596830, figure 2A}
\format{
A data frame of 681 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (control, figitumumab) \cr
}
}
\source{
Langer CJ, Novello S, Park K, et al. Randomized, phase III trial of
first-line figitumumab in combination with paclitaxel and carboplatin
versus paclitaxel and carboplatin alone in patients with advanced
non-small-cell lung cancer. J Clin Oncol 2014; 32: 2059–66.
}
\usage{
NCT00596830_2A
}
\description{
Kaplan-Meier digitized data from NCT00596830, figure 2A (PMID 24888810). A reported sample size of 681 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(NCT00596830_2A)

kmplot(NCT00596830_2A)
}
\keyword{datasets}
