% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCT00614393(a)_3A}
\alias{NCT00614393(a)_3A}
\title{NCT00614393(a), figure 3A}
\format{
A data frame of 227 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (arma, armc) \cr
}
}
\source{
Sclafani F, Kim TY, Cunningham D, et al. A Randomized Phase II/III
Study of Dalotuzumab in Combination With Cetuximab and Irinotecan in
Chemorefractory, KRAS Wild-Type, Metastatic Colorectal Cancer. J Natl
Cancer Inst 2015; 107: djv258.
}
\usage{
`NCT00614393(a)_3A`
}
\description{
Kaplan-Meier digitized data from NCT00614393(a), figure 3A (PMID 26405092). A reported sample size of 344 for a primary endpoint of PFS,OS in colorectal cancer.
}
\examples{
summary(`NCT00614393(a)_3A`)

kmplot(`NCT00614393(a)_3A`)
}
\keyword{datasets}
