% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCT00673049_2B}
\alias{NCT00673049_2B}
\title{NCT00673049, figure 2B}
\format{
A data frame of 583 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (control, figitumumab) \cr
}
}
\source{
Scagliotti GV, Bondarenko I, Blackhall F, et al. Randomized, phase
III trial of figitumumab in combination with erlotinib versus
erlotinib alone in patients with nonadenocarcinoma nonsmall-cell lung
cancer. Ann Oncol 2015; 26: 497–504.
}
\usage{
NCT00673049_2B
}
\description{
Kaplan-Meier digitized data from NCT00673049, figure 2B (PMID 25395283). A reported sample size of 583 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(NCT00673049_2B)

kmplot(NCT00673049_2B)
}
\keyword{datasets}
