% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCT00938652_2C}
\alias{NCT00938652_2C}
\title{NCT00938652, figure 2C}
\format{
A data frame of 519 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (gc, gci) \cr
}
}
\source{
O’Shaughnessy J, Schwartzberg L, Danso MA, et al. Phase III study of
iniparib plus gemcitabine and carboplatin versus gemcitabine and
carboplatin in patients with metastatic triple-negative breast
cancer. J Clin Oncol 2014; 32: 3840–7.
}
\usage{
NCT00938652_2C
}
\description{
Kaplan-Meier digitized data from NCT00938652, figure 2C (PMID 25349301). A reported sample size of 519 for a primary endpoint of OS/PFS in breast cancer.
}
\examples{
summary(NCT00938652_2C)

kmplot(NCT00938652_2C)
}
\keyword{datasets}
