% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCT01234311_2A}
\alias{NCT01234311_2A}
\title{NCT01234311, figure 2A}
\format{
A data frame of 577 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (placebo, tasq) \cr
}
}
\source{
Sternberg C, Armstrong A, Pili R, et al. Randomized, Double-Blind,
Placebo-Controlled Phase III Study of Tasquinimod in Men With
Metastatic Castration-Resistant Prostate Cancer. J Clin Oncol 2016;
34: 2636–43.
}
\usage{
NCT01234311_2A
}
\description{
Kaplan-Meier digitized data from NCT01234311, figure 2A (PMID 27298414). A reported sample size of 1,245 for a primary endpoint of rPFS in prostate cancer.
}
\examples{
summary(NCT01234311_2A)

kmplot(NCT01234311_2A)
}
\keyword{datasets}
