% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NeoALTTO(a)_2A}
\alias{NeoALTTO(a)_2A}
\title{NeoALTTO(a), figure 2A}
\format{
A data frame of 303 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab EFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (l, t) \cr
}
}
\source{
de Azambuja E, Holmes AP, Piccart-Gebhart M, et al. Lapatinib with
trastuzumab for HER2-positive early breast cancer (NeoALTTO):
survival outcomes of a randomised, open-label, multicentre, phase 3
trial and their association with pathological complete response.
Lancet Oncol 2014; 15: 1137–46.
}
\usage{
`NeoALTTO(a)_2A`
}
\description{
Kaplan-Meier digitized data from NeoALTTO(a), figure 2A (PMID 25130998). A reported sample size of 455 for a primary endpoint of pCR in breast cancer.
}
\examples{
summary(`NeoALTTO(a)_2A`)

kmplot(`NeoALTTO(a)_2A`)
}
\keyword{datasets}
