% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NITRO_2B}
\alias{NITRO_2B}
\title{NITRO, figure 2B}
\format{
A data frame of 372 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (chemo, nitro_chemo) \cr
}
}
\source{
Davidson A, Veillard A-S, Tognela A, et al. A phase III randomized
trial of adding topical nitroglycerin to first-line chemotherapy for
advanced nonsmall-cell lung cancer: the Australasian lung cancer
trials group NITRO trial. Ann Oncol 2015; 26: 2280–6.
}
\usage{
NITRO_2B
}
\description{
Kaplan-Meier digitized data from NITRO, figure 2B (PMID 26347110). A reported sample size of 372 for a primary endpoint of PFS in lung cancer.
}
\examples{
summary(NITRO_2B)

kmplot(NITRO_2B)
}
\keyword{datasets}
