% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NO16968_2C}
\alias{NO16968_2C}
\title{NO16968, figure 2C}
\format{
A data frame of 1,886 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (fu_fa, xelox) \cr
}
}
\source{
Schmoll H-J, Tabernero J, Maroun J, et al. Capecitabine Plus
Oxaliplatin Compared With Fluorouracil/Folinic Acid As Adjuvant
Therapy for Stage III Colon Cancer: Final Results of the NO16968
Randomized Controlled Phase III Trial. J Clin Oncol 2015; 33:
3733–40.
}
\usage{
NO16968_2C
}
\description{
Kaplan-Meier digitized data from NO16968, figure 2C (PMID 26324362). A reported sample size of 1,886 for a primary endpoint of DFS in colorectal cancer.
}
\examples{
summary(NO16968_2C)

kmplot(NO16968_2C)
}
\keyword{datasets}
