% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOAH_2B}
\alias{NOAH_2B}
\title{NOAH, figure 2B}
\format{
A data frame of 235 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (chemo, chemotherapy_trastuzumab) \cr
}
}
\source{
Gianni L, Eiermann W, Semiglazov V, et al. Neoadjuvant and adjuvant
trastuzumab in patients with HER2-positive locally advanced breast
cancer (NOAH): follow-up of a randomised controlled superiority trial
with a parallel HER2-negative cohort. Lancet Oncol 2014; 15: 640–7.
}
\usage{
NOAH_2B
}
\description{
Kaplan-Meier digitized data from NOAH, figure 2B (PMID 24657003). A reported sample size of 235 for a primary endpoint of EFS in breast cancer.
}
\examples{
summary(NOAH_2B)

kmplot(NOAH_2B)
}
\keyword{datasets}
