% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NSABPB35_3}
\alias{NSABPB35_3}
\title{NSABPB35, figure 3}
\format{
A data frame of 3,077 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (anastrazole, tamoxifen) \cr
}
}
\source{
Margolese RG, Cecchini RS, Julian TB, et al. Anastrozole versus
tamoxifen in postmenopausal women with ductal carcinoma in situ
undergoing lumpectomy plus radiotherapy (NSABP B-35): a randomised,
double-blind, phase 3 clinical trial. Lancet 2016; 387: 849–56.
}
\usage{
NSABPB35_3
}
\description{
Kaplan-Meier digitized data from NSABPB35, figure 3 (PMID 26686957). A reported sample size of 3,104 for a primary endpoint of BCFS in breast cancer.
}
\examples{
summary(NSABPB35_3)

kmplot(NSABPB35_3)
}
\keyword{datasets}
