% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NSABPB40_3A}
\alias{NSABPB40_3A}
\title{NSABPB40, figure 3A}
\format{
A data frame of 1,184 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (bev, no_bev) \cr
}
}
\source{
Bear HD, Tang G, Rastogi P, et al. Neoadjuvant plus adjuvant
bevacizumab in early breast cancer (NSABP B-40 [NRG Oncology]):
secondary outcomes of a phase 3, randomised controlled trial. Lancet
Oncol 2015; 16: 1037–48.
}
\usage{
NSABPB40_3A
}
\description{
Kaplan-Meier digitized data from NSABPB40, figure 3A (PMID 26272770). A reported sample size of 1,206 for a primary endpoint of pCR in breast cancer.
}
\examples{
summary(NSABPB40_3A)

kmplot(NSABPB40_3A)
}
\keyword{datasets}
