% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NSABPR04_3B}
\alias{NSABPR04_3B}
\title{NSABPR04, figure 3B}
\format{
A data frame of 1,284 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab LRR event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (no_oxali, oxali) \cr
}
}
\source{
Allegra CJ, Yothers G, O’Connell MJ, et al. Neoadjuvant 5-FU or
Capecitabine Plus Radiation With or Without Oxaliplatin in Rectal
Cancer Patients: A Phase III Randomized Clinical Trial. J Natl Cancer
Inst 2015; 107. DOI:10.1093/jnci/djv248.
}
\usage{
NSABPR04_3B
}
\description{
Kaplan-Meier digitized data from NSABPR04, figure 3B (PMID 26374429). A reported sample size of 1,608 for a primary endpoint of Local-regional tumor control in colorectal cancer.
}
\examples{
summary(NSABPR04_3B)

kmplot(NSABPR04_3B)
}
\keyword{datasets}
