% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NTR1527_4}
\alias{NTR1527_4}
\title{NTR1527, figure 4}
\format{
A data frame of 495 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (control, thoracic rt) \cr
}
}
\source{
Slotman BJ, van Tinteren H, Praag JO, et al. Use of thoracic
radiotherapy for extensive stage small-cell lung cancer: a phase 3
randomised controlled trial. Lancet 2015; 385: 36–42.
}
\usage{
NTR1527_4
}
\description{
Kaplan-Meier digitized data from NTR1527, figure 4 (PMID 25230595). A reported sample size of 498 for a primary endpoint of OS-1yr in lung cancer.
}
\examples{
summary(NTR1527_4)

kmplot(NTR1527_4)
}
\keyword{datasets}
