% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OMEGA_3B}
\alias{OMEGA_3B}
\title{OMEGA, figure 3B}
\format{
A data frame of 78 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (cap, peg_doxorubicin) \cr
}
}
\source{
Smorenburg CH, de Groot SM, van Leeuwen-Stok AE, et al. A randomized
phase III study comparing pegylated liposomal doxorubicin with
capecitabine as first-line chemotherapy in elderly patients with
metastatic breast cancer: results of the OMEGA study of the Dutch
Breast Cancer Research Group BOOG. Ann Oncol 2014; 25: 599–605.
}
\usage{
OMEGA_3B
}
\description{
Kaplan-Meier digitized data from OMEGA, figure 3B (PMID 24504445). A reported sample size of 78 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(OMEGA_3B)

kmplot(OMEGA_3B)
}
\keyword{datasets}
