% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OPTIMAL_1}
\alias{OPTIMAL_1}
\title{OPTIMAL, figure 1}
\format{
A data frame of 154 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (erlotinib, gem_carboplatin) \cr
}
}
\source{
Zhou C, Wu YL, Chen G, et al. Final overall survival results from a
randomised, phase III study of erlotinib versus chemotherapy as
first-line treatment of EGFR mutation-positive advanced
non-small-cell lung cancer (OPTIMAL, CTONG-0802). Ann Oncol 2015; 26:
1877–83.
}
\usage{
OPTIMAL_1
}
\description{
Kaplan-Meier digitized data from OPTIMAL, figure 1 (PMID 26141208). A reported sample size of 165 for a primary endpoint of PFS in lung cancer.
}
\examples{
summary(OPTIMAL_1)

kmplot(OPTIMAL_1)
}
\keyword{datasets}
