% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OPTIMOX3_3B}
\alias{OPTIMOX3_3B}
\title{OPTIMOX3, figure 3B}
\format{
A data frame of 452 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (bev, bev_erlotinib) \cr
}
}
\source{
Tournigand C, Chibaudel B, Samson B, et al. Bevacizumab with or
without erlotinib as maintenance therapy in patients with metastatic
colorectal cancer (GERCOR DREAM; OPTIMOX3): a randomised, open-label,
phase 3 trial. Lancet Oncol 2015; 16: 1493–505.
}
\usage{
OPTIMOX3_3B
}
\description{
Kaplan-Meier digitized data from OPTIMOX3, figure 3B (PMID 26474518). A reported sample size of 700 for a primary endpoint of PFS in colorectal cancer.
}
\examples{
summary(OPTIMOX3_3B)

kmplot(OPTIMOX3_3B)
}
\keyword{datasets}
