% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PALOMA2_1A}
\alias{PALOMA2_1A}
\title{PALOMA2, figure 1A}
\format{
A data frame of 666 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (palbociclib_letrozole, placebo_letrozole) \cr
}
}
\source{
Finn RS, Martin M, Rugo HS, et al. Palbociclib and Letrozole in
Advanced Breast Cancer. N Engl J Med 2016; 375: 1925–36.
}
\usage{
PALOMA2_1A
}
\description{
Kaplan-Meier digitized data from PALOMA2, figure 1A (PMID 27959613). A reported sample size of 666 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(PALOMA2_1A)

kmplot(PALOMA2_1A)
}
\keyword{datasets}
