% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PALOMA3_1A}
\alias{PALOMA3_1A}
\title{PALOMA3, figure 1A}
\format{
A data frame of 521 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (palbociclib_fulvestrant, placebo_fulvestrant) \cr
}
}
\source{
Turner NC, Ro J, André F, et al. Palbociclib in
Hormone-Receptor-Positive Advanced Breast Cancer. N Engl J Med 2015;
373: 209–19.
}
\usage{
PALOMA3_1A
}
\description{
Kaplan-Meier digitized data from PALOMA3, figure 1A (PMID 26030518). A reported sample size of 521 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(PALOMA3_1A)

kmplot(PALOMA3_1A)
}
\keyword{datasets}
