% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PALOMA3_2A}
\alias{PALOMA3_2A}
\title{PALOMA3, figure 2A}
\format{
A data frame of 521 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (fulvestrant_palbociclib, fulvestrant_placebo) \cr
}
}
\source{
Cristofanilli M, Turner NC, Bondarenko I, et al. Fulvestrant plus
palbociclib versus fulvestrant plus placebo for treatment of
hormone-receptor-positive, HER2-negative metastatic breast cancer
that progressed on previous endocrine therapy (PALOMA-3): final
analysis of the multicentre, double-blind, phase 3 randomised
controlled trial. Lancet Oncol 2016; 17: 425–39.
}
\usage{
PALOMA3_2A
}
\description{
Kaplan-Meier digitized data from PALOMA3, figure 2A (PMID 26947331). A reported sample size of 521 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(PALOMA3_2A)

kmplot(PALOMA3_2A)
}
\keyword{datasets}
