% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PEGASE07_2B}
\alias{PEGASE07_2B}
\title{PEGASE07, figure 2B}
\format{
A data frame of 174 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (arma, armb) \cr
}
}
\source{
Gonçalves A, Pierga J-Y, Ferrero J-M, et al. UNICANCER-PEGASE 07
study: a randomized phase III trial evaluating postoperative
docetaxel-5FU regimen after neoadjuvant dose-intense chemotherapy for
treatment of inflammatory breast cancer. Ann Oncol 2015; 26: 1692–7.
}
\usage{
PEGASE07_2B
}
\description{
Kaplan-Meier digitized data from PEGASE07, figure 2B (PMID 25943350). A reported sample size of 174 for a primary endpoint of DFS in breast cancer.
}
\examples{
summary(PEGASE07_2B)

kmplot(PEGASE07_2B)
}
\keyword{datasets}
