% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PREVAIL_1A}
\alias{PREVAIL_1A}
\title{PREVAIL, figure 1A}
\format{
A data frame of 1,633 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab RPFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (enzalutamide, placebo) \cr
}
}
\source{
Beer TM, Armstrong AJ, Rathkopf DE, et al. Enzalutamide in metastatic
prostate cancer before chemotherapy. N Engl J Med 2014; 371: 424–33.
}
\usage{
PREVAIL_1A
}
\description{
Kaplan-Meier digitized data from PREVAIL, figure 1A (PMID 24881730). A reported sample size of 1,717 for a primary endpoint of PFS/OS in prostate cancer.
}
\examples{
summary(PREVAIL_1A)

kmplot(PREVAIL_1A)
}
\keyword{datasets}
