% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROCLAIM_2A}
\alias{PROCLAIM_2A}
\title{PROCLAIM, figure 2A}
\format{
A data frame of 598 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (eto_cis, pem_cis) \cr
}
}
\source{
Senan S, Brade A, Wang L-H, et al. PROCLAIM: Randomized Phase III
Trial of Pemetrexed-Cisplatin or Etoposide-Cisplatin Plus Thoracic
Radiation Therapy Followed by Consolidation Chemotherapy in Locally
Advanced Nonsquamous Non-Small-Cell Lung Cancer. J Clin Oncol 2016;
34: 953–62.
}
\usage{
PROCLAIM_2A
}
\description{
Kaplan-Meier digitized data from PROCLAIM, figure 2A (PMID 26811519). A reported sample size of 598 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(PROCLAIM_2A)

kmplot(PROCLAIM_2A)
}
\keyword{datasets}
