% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROCTOR-SCRIPT_1A}
\alias{PROCTOR-SCRIPT_1A}
\title{PROCTOR-SCRIPT, figure 1A}
\format{
A data frame of 437 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (chemo, observation) \cr
}
}
\source{
Breugom AJ, van Gijn W, Muller EW, et al. Adjuvant chemotherapy for
rectal cancer patients treated with preoperative (chemo)radiotherapy
and total mesorectal excision: a Dutch Colorectal Cancer Group (DCCG)
randomized phase III trial. Ann Oncol 2015; 26: 696–701.
}
\usage{
`PROCTOR-SCRIPT_1A`
}
\description{
Kaplan-Meier digitized data from PROCTOR-SCRIPT, figure 1A (PMID 25480874). A reported sample size of 437 for a primary endpoint of OS in colorectal cancer.
}
\examples{
summary(`PROCTOR-SCRIPT_1A`)

kmplot(`PROCTOR-SCRIPT_1A`)
}
\keyword{datasets}
