% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROFILE1014_1A}
\alias{PROFILE1014_1A}
\title{PROFILE1014, figure 1A}
\format{
A data frame of 343 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (chemo, crizotinib) \cr
}
}
\source{
Solomon BJ, Mok T, Kim D-W, et al. First-line crizotinib versus
chemotherapy in ALK-positive lung cancer. N Engl J Med 2014; 371:
2167–77.
}
\usage{
PROFILE1014_1A
}
\description{
Kaplan-Meier digitized data from PROFILE1014, figure 1A (PMID 25470694). A reported sample size of 343 for a primary endpoint of PFS in lung cancer.
}
\examples{
summary(PROFILE1014_1A)

kmplot(PROFILE1014_1A)
}
\keyword{datasets}
