% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROFILE1014_2A}
\alias{PROFILE1014_2A}
\title{PROFILE1014, figure 2A}
\format{
A data frame of 343 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (chemo, crizotinib) \cr
}
}
\source{
Solomon BJ, Cappuzzo F, Felip E, et al. Intracranial Efficacy of
Crizotinib Versus Chemotherapy in Patients With Advanced ALK-Positive
Non–Small-Cell Lung Cancer: Results From PROFILE 1014. J Clin Oncol
2016; 34: 2858-67.
}
\usage{
PROFILE1014_2A
}
\description{
Kaplan-Meier digitized data from PROFILE1014, figure 2A (PMID 27022118). A reported sample size of 343 for a primary endpoint of PFS in lung cancer.
}
\examples{
summary(PROFILE1014_2A)

kmplot(PROFILE1014_2A)
}
\keyword{datasets}
