% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROSE_2A}
\alias{PROSE_2A}
\title{PROSE, figure 2A}
\format{
A data frame of 263 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (chemo, erlotinib) \cr
}
}
\source{
Gregorc V, Novello S, Lazzari C, et al. Predictive value of a
proteomic signature in patients with non-small-cell lung cancer
treated with second-line erlotinib or chemotherapy (PROSE): a
biomarker-stratified, randomised phase 3 trial. Lancet Oncol 2014;
15: 713–21.
}
\usage{
PROSE_2A
}
\description{
Kaplan-Meier digitized data from PROSE, figure 2A (PMID 24831979). A reported sample size of 142 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(PROSE_2A)

kmplot(PROSE_2A)
}
\keyword{datasets}
