% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QUARTZ_2A}
\alias{QUARTZ_2A}
\title{QUARTZ, figure 2A}
\format{
A data frame of 538 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in weeks) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (osc, osc_wbrt) \cr
}
}
\source{
Mulvenna P, Nankivell M, Barton R, et al. Dexamethasone and
supportive care with or without whole brain radiotherapy in treating
patients with non-small cell lung cancer with brain metastases
unsuitable for resection or stereotactic radiotherapy (QUARTZ):
results from a phase 3, non-inferiority, randomised trial. Lancet
2016; 388: 2004–14.
}
\usage{
QUARTZ_2A
}
\description{
Kaplan-Meier digitized data from QUARTZ, figure 2A (PMID 27604504). A reported sample size of 538 for a primary endpoint of QALYs in lung cancer.
}
\examples{
summary(QUARTZ_2A)

kmplot(QUARTZ_2A)
}
\keyword{datasets}
