% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QUASAR2_2A}
\alias{QUASAR2_2A}
\title{QUASAR2, figure 2A}
\format{
A data frame of 1,933 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (cap, cap_bev) \cr
}
}
\source{
Kerr RS, Love S, Segelov E, et al. Adjuvant capecitabine plus
bevacizumab versus capecitabine alone in patients with colorectal
cancer (QUASAR 2): an open-label, randomised phase 3 trial. Lancet
Oncol 2016; 17: 1543–57.
}
\usage{
QUASAR2_2A
}
\description{
Kaplan-Meier digitized data from QUASAR2, figure 2A (PMID 27660192). A reported sample size of 1,952 for a primary endpoint of DFS3 in colorectal cancer.
}
\examples{
summary(QUASAR2_2A)

kmplot(QUASAR2_2A)
}
\keyword{datasets}
