% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R98_4}
\alias{R98_4}
\title{R98, figure 4}
\format{
A data frame of 357 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (5fulv, 5fulv_cpt11) \cr
}
}
\source{
Delbaldo C, Ychou M, Zawadi A, et al. Postoperative irinotecan in
resected stage II-III rectal cancer: final analysis of the French R98
Intergroup trial†. Ann Oncol 2015; 26: 1208–15.
}
\usage{
R98_4
}
\description{
Kaplan-Meier digitized data from R98, figure 4 (PMID 25739671). A reported sample size of 357 for a primary endpoint of DFS in colorectal cancer.
}
\examples{
summary(R98_4)

kmplot(R98_4)
}
\keyword{datasets}
