% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RADIANT4_2C}
\alias{RADIANT4_2C}
\title{RADIANT4, figure 2C}
\format{
A data frame of 302 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (everolimus, placebo) \cr
}
}
\source{
Yao JC, Fazio N, Singh S, et al. Everolimus for the treatment of
advanced, non-functional neuroendocrine tumours of the lung or
gastrointestinal tract (RADIANT-4): a randomised, placebo-controlled,
phase 3 study. Lancet 2016; 387: 968–77.
}
\usage{
RADIANT4_2C
}
\description{
Kaplan-Meier digitized data from RADIANT4, figure 2C (PMID 26703889). A reported sample size of 302 for a primary endpoint of PFS in lung/colorectal cancer.
}
\examples{
summary(RADIANT4_2C)

kmplot(RADIANT4_2C)
}
\keyword{datasets}
