% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RADIANT_2A}
\alias{RADIANT_2A}
\title{RADIANT, figure 2A}
\format{
A data frame of 973 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (erlotinib, placebo) \cr
}
}
\source{
Kelly K, Altorki NK, Eberhardt WEE, et al. Adjuvant Erlotinib Versus
Placebo in Patients With Stage IB-IIIA Non-Small-Cell Lung Cancer
(RADIANT): A Randomized, Double-Blind, Phase III Trial. J Clin Oncol
2015; 33: 4007–14.
}
\usage{
RADIANT_2A
}
\description{
Kaplan-Meier digitized data from RADIANT, figure 2A (PMID 26324372). A reported sample size of 973 for a primary endpoint of DFS in lung cancer.
}
\examples{
summary(RADIANT_2A)

kmplot(RADIANT_2A)
}
\keyword{datasets}
