% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RAISE_2A}
\alias{RAISE_2A}
\title{RAISE, figure 2A}
\format{
A data frame of 1,072 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (placebo_folfiri, ramucirumab_folfiri) \cr
}
}
\source{
Tabernero J, Yoshino T, Cohn AL, et al. Ramucirumab versus placebo in
combination with second-line FOLFIRI in patients with metastatic
colorectal carcinoma that progressed during or after first-line
therapy with bevacizumab, oxaliplatin, and a fluoropyrimidine
(RAISE): a randomised, double-blind, multicentre, phase 3 study.
Lancet Oncol 2015; 16: 499–508.
}
\usage{
RAISE_2A
}
\description{
Kaplan-Meier digitized data from RAISE, figure 2A (PMID 25877855). A reported sample size of 1,072 for a primary endpoint of OS in colorectal cancer.
}
\examples{
summary(RAISE_2A)

kmplot(RAISE_2A)
}
\keyword{datasets}
