% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RECOURSE_1A}
\alias{RECOURSE_1A}
\title{RECOURSE, figure 1A}
\format{
A data frame of 800 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (placebo, tas-102) \cr
}
}
\source{
Mayer RJ, Van Cutsem E, Falcone A, et al. Randomized trial of TAS-102
for refractory metastatic colorectal cancer. N Engl J Med 2015; 372:
1909–19.
}
\usage{
RECOURSE_1A
}
\description{
Kaplan-Meier digitized data from RECOURSE, figure 1A (PMID 25970050). A reported sample size of 800 for a primary endpoint of OS in breast cancer.
}
\examples{
summary(RECOURSE_1A)

kmplot(RECOURSE_1A)
}
\keyword{datasets}
