% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REVEL_3}
\alias{REVEL_3}
\title{REVEL, figure 3}
\format{
A data frame of 1,253 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (placebo_docetaxel, ramucirumab_docetaxel) \cr
}
}
\source{
Garon EB, Ciuleanu T-E, Arrieta O, et al. Ramucirumab plus docetaxel
versus placebo plus docetaxel for second-line treatment of stage IV
non-small-cell lung cancer after disease progression on
platinum-based therapy (REVEL): a multicentre, double-blind,
randomised phase 3 trial. Lancet 2014; 384: 665–73.
}
\usage{
REVEL_3
}
\description{
Kaplan-Meier digitized data from REVEL, figure 3 (PMID 24933332). A reported sample size of 1,825 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(REVEL_3)

kmplot(REVEL_3)
}
\keyword{datasets}
