% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ROSE_2C}
\alias{ROSE_2C}
\title{ROSE, figure 2C}
\format{
A data frame of 1,144 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (pbo_doc, ram_doc) \cr
}
}
\source{
Mackey JR, Ramos-Vazquez M, Lipatov O, et al. Primary results of
ROSE/TRIO-12, a randomized placebo-controlled phase III trial
evaluating the addition of ramucirumab to first-line docetaxel
chemotherapy in metastatic breast cancer. J Clin Oncol 2015; 33:
141–8.
}
\usage{
ROSE_2C
}
\description{
Kaplan-Meier digitized data from ROSE, figure 2C (PMID 25185099). A reported sample size of 1,144 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(ROSE_2C)

kmplot(ROSE_2C)
}
\keyword{datasets}
