% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RTOG0415_2}
\alias{RTOG0415_2}
\title{RTOG0415, figure 2}
\format{
A data frame of 1,092 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (70gy, 73.8gy) \cr
}
}
\source{
Lee WR, Dignam JJ, Amin MB, et al. Randomized Phase III
Noninferiority Study Comparing Two Radiotherapy Fractionation
Schedules in Patients With Low-Risk Prostate Cancer. J Clin Oncol
2016; 34: 2325–32.
}
\usage{
RTOG0415_2
}
\description{
Kaplan-Meier digitized data from RTOG0415, figure 2 (PMID 27044935). A reported sample size of 1,115 for a primary endpoint of DFS in prostate cancer.
}
\examples{
summary(RTOG0415_2)

kmplot(RTOG0415_2)
}
\keyword{datasets}
