% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RTOG0617_2B}
\alias{RTOG0617_2B}
\title{RTOG0617, figure 2B}
\format{
A data frame of 465 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (cetuximab, no_cetuximab) \cr
}
}
\source{
Bradley JD, Paulus R, Komaki R, et al. Standard-dose versus high-dose
conformal radiotherapy with concurrent and consolidation carboplatin
plus paclitaxel with or without cetuximab for patients with stage
IIIA or IIIB non-small-cell lung cancer (RTOG 0617): a randomised,
two-by-two factorial phase 3 study. Lancet Oncol 2015; 16: 187–99.
}
\usage{
RTOG0617_2B
}
\description{
Kaplan-Meier digitized data from RTOG0617, figure 2B (PMID 25601342). A reported sample size of 166 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(RTOG0617_2B)

kmplot(RTOG0617_2B)
}
\keyword{datasets}
