% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RTOG9804_2D}
\alias{RTOG9804_2D}
\title{RTOG9804, figure 2D}
\format{
A data frame of 585 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (observation, rt) \cr
}
}
\source{
McCormick B, Winter K, Hudis C, et al. RTOG 9804: a prospective
randomized trial for good-risk ductal carcinoma in situ comparing
radiotherapy with observation. J Clin Oncol 2015; 33: 709–15.
}
\usage{
RTOG9804_2D
}
\description{
Kaplan-Meier digitized data from RTOG9804, figure 2D (PMID 25605856). A reported sample size of 632 for a primary endpoint of ILF in breast cancer.
}
\examples{
summary(RTOG9804_2D)

kmplot(RTOG9804_2D)
}
\keyword{datasets}
