% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SAKK4106_1}
\alias{SAKK4106_1}
\title{SAKK4106, figure 1}
\format{
A data frame of 262 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (bev, no_bev) \cr
}
}
\source{
Koeberle D, Betticher DC, von Moos R, et al. Bevacizumab continuation
versus no continuation after first-line chemotherapy plus bevacizumab
in patients with metastatic colorectal cancer: a randomized phase III
non-inferiority trial (SAKK 41/06). Ann Oncol 2015; 26: 709–14.
}
\usage{
SAKK4106_1
}
\description{
Kaplan-Meier digitized data from SAKK4106, figure 1 (PMID 25605741). A reported sample size of 262 for a primary endpoint of TTP in colorectal cancer.
}
\examples{
summary(SAKK4106_1)

kmplot(SAKK4106_1)
}
\keyword{datasets}
