% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SAKK_2B}
\alias{SAKK_2B}
\title{SAKK, figure 2B}
\format{
A data frame of 232 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (chemo_surgery, crt_surgery) \cr
}
}
\source{
Pless M, Stupp R, Ris H-B, et al. Induction chemoradiation in stage
IIIA/N2 non-small-cell lung cancer: a phase 3 randomised trial.
Lancet 2015; 386: 1049–56.
}
\usage{
SAKK_2B
}
\description{
Kaplan-Meier digitized data from SAKK, figure 2B (PMID 26275735). A reported sample size of 232 for a primary endpoint of EFS in lung cancer.
}
\examples{
summary(SAKK_2B)

kmplot(SAKK_2B)
}
\keyword{datasets}
