% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmdata-package.R
\name{select_kmdata}
\alias{select_kmdata}
\title{Select data sets}
\usage{
select_kmdata(..., return = c("name", "key", "data"))
}
\arguments{
\item{...}{an expression to be evaluated within \code{\link{kmdata_key}}
such as \code{SampleSize < 100}; see examples}

\item{return}{type of object to return; one of \code{"name"} (default) for
the study names that match the criteria, \code{"key"} for the matching
rows of \code{kmdata_key}, or \code{"data"} for a list of data frames for
each match}
}
\description{
Select publication data sets based on study characteristics including
outcome, sample size, treatment arms, journal, disease, etc.
}
\examples{
names(kmdata_key)
select_kmdata(SampleSize < 100)
select_kmdata(grepl('folfiri', Arms) & Outcome == 'OS')
select_kmdata(SampleSize < 100 | Cancer \%in\% c('Lung/Colorectal', 'Prostate'))

## get a list of the data sets
l <- select_kmdata(SampleSize < 100, return = 'data')
par(mfrow = n2mfrow(length(l)))
sapply(l, kmplot)

}
