% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SELECTBC_2B}
\alias{SELECTBC_2B}
\title{SELECTBC, figure 2B}
\format{
A data frame of 592 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab TTF event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (s1, taxane) \cr
}
}
\source{
Takashima T, Mukai H, Hara F, et al. Taxanes versus S-1 as the
first-line chemotherapy for metastatic breast cancer (SELECT BC): an
open-label, non-inferiority, randomised phase 3 trial. Lancet Oncol
2016; 17: 90–8.
}
\usage{
SELECTBC_2B
}
\description{
Kaplan-Meier digitized data from SELECTBC, figure 2B (PMID 26617202). A reported sample size of 618 for a primary endpoint of OS in breast cancer.
}
\examples{
summary(SELECTBC_2B)

kmplot(SELECTBC_2B)
}
\keyword{datasets}
