% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SIRFLOX_3}
\alias{SIRFLOX_3}
\title{SIRFLOX, figure 3}
\format{
A data frame of 530 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (mfolfox6, mfolfox6_rt) \cr
}
}
\source{
van Hazel GA, Heinemann V, Sharma NK, et al. SIRFLOX: Randomized
Phase III Trial Comparing First-Line mFOLFOX6 (Plus or Minus
Bevacizumab) Versus mFOLFOX6 (Plus or Minus Bevacizumab) Plus
Selective Internal Radiation Therapy in Patients With Metastatic
Colorectal Cancer. J Clin Oncol 2016; 34: 1723–31.
}
\usage{
SIRFLOX_3
}
\description{
Kaplan-Meier digitized data from SIRFLOX, figure 3 (PMID 26903575). A reported sample size of 530 for a primary endpoint of PFS in colorectal cancer.
}
\examples{
summary(SIRFLOX_3)

kmplot(SIRFLOX_3)
}
\keyword{datasets}
