% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SMART_3}
\alias{SMART_3}
\title{SMART, figure 3}
\format{
A data frame of 203 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in days) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (deferred_rt, immediate_rt) \cr
}
}
\source{
Clive AO, Taylor H, Dobson L, et al. Prophylactic radiotherapy for
the prevention of procedure-tract metastases after surgical and
large-bore pleural procedures in malignant pleural mesothelioma
(SMART): a multicentre, open-label, phase 3, randomised controlled
trial. Lancet Oncol 2016; 17: 1094–104.
}
\usage{
SMART_3
}
\description{
Kaplan-Meier digitized data from SMART, figure 3 (PMID 27345639). A reported sample size of 203 for a primary endpoint of PTM-incidence in lung cancer.
}
\examples{
summary(SMART_3)

kmplot(SMART_3)
}
\keyword{datasets}
