% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SMX1132531_1}
\alias{SMX1132531_1}
\title{SMX1132531, figure 1}
\format{
A data frame of 98 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (srs, upfront_chemotherapy) \cr
}
}
\source{
Lim SH, Lee JY, Lee M-Y, et al. A randomized phase III trial of
stereotactic radiosurgery (SRS) versus observation for patients with
asymptomatic cerebral oligo-metastases in non-small-cell lung cancer.
Ann Oncol 2015; 26: 762–8.
}
\usage{
SMX1132531_1
}
\description{
Kaplan-Meier digitized data from SMX1132531, figure 1 (PMID 25538174). A reported sample size of 105 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(SMX1132531_1)

kmplot(SMX1132531_1)
}
\keyword{datasets}
