% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SOFT_2A}
\alias{SOFT_2A}
\title{SOFT, figure 2A}
\format{
A data frame of 2,033 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (tamoxifen, tamoxifen_os) \cr
}
}
\source{
Francis PA, Regan MM, Fleming GF, et al. Adjuvant ovarian suppression
in premenopausal breast cancer. N Engl J Med 2015; 372: 436–46.
}
\usage{
SOFT_2A
}
\description{
Kaplan-Meier digitized data from SOFT, figure 2A (PMID 25495490). A reported sample size of 3,066 for a primary endpoint of DFS in breast cancer.
}
\examples{
summary(SOFT_2A)

kmplot(SOFT_2A)
}
\keyword{datasets}
