% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmdata-package.R
\name{summary.kmdata}
\alias{summary.kmdata}
\title{Summarize \code{kmdata}}
\usage{
\method{summary}{kmdata}(object, message = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{"kmdata"}}

\item{message}{logical; if \code{TRUE} (default), a description of the
study will be printed with the summary}

\item{...}{ignored}
}
\description{
Print a summary of a \code{kmdata} object.
}
\examples{
summary(ATTENTION_2A)

}
