% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TANIA_2}
\alias{TANIA_2}
\title{TANIA, figure 2}
\format{
A data frame of 494 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (chem_bev, chemo) \cr
}
}
\source{
von Minckwitz G, Puglisi F, Cortes J, et al. Bevacizumab plus
chemotherapy versus chemotherapy alone as second-line treatment for
patients with HER2-negative locally recurrent or metastatic breast
cancer after first-line treatment with bevacizumab plus chemotherapy
(TANIA): an open-label, randomised phase 3 trial. Lancet Oncol 2014;
15: 1269–78.
}
\usage{
TANIA_2
}
\description{
Kaplan-Meier digitized data from TANIA, figure 2 (PMID 25273342). A reported sample size of 494 for a primary endpoint of PFS in breast cancer.
}
\examples{
summary(TANIA_2)

kmplot(TANIA_2)
}
\keyword{datasets}
