% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCOG0701_1B}
\alias{TCOG0701_1B}
\title{TCOG0701, figure 1B}
\format{
A data frame of 596 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (docetaxel_cis, s1_cis) \cr
}
}
\source{
Kubota K, Sakai H, Katakami N, et al. A randomized phase III trial of
oral S-1 plus cisplatin versus docetaxel plus cisplatin in Japanese
patients with advanced non-small-cell lung cancer: TCOG0701 CATS
trial. Ann Oncol 2015; 26: 1401–8.
}
\usage{
TCOG0701_1B
}
\description{
Kaplan-Meier digitized data from TCOG0701, figure 1B (PMID 25908605). A reported sample size of 608 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(TCOG0701_1B)

kmplot(TCOG0701_1B)
}
\keyword{datasets}
