% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TEXT-SOFT(2)_2D}
\alias{TEXT-SOFT(2)_2D}
\title{TEXT-SOFT(2), figure 2D}
\format{
A data frame of 4,690 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (exemestane_os, tamoxifen_os) \cr
}
}
\source{
Pagani O, Regan MM, Walley BA, et al. Adjuvant exemestane with
ovarian suppression in premenopausal breast cancer. N Engl J Med
2014; 371: 107–18.
}
\usage{
`TEXT-SOFT(2)_2D`
}
\description{
Kaplan-Meier digitized data from TEXT-SOFT(2), figure 2D (PMID 24881463). A reported sample size of 4,690 for a primary endpoint of DFS in breast cancer.
}
\examples{
summary(`TEXT-SOFT(2)_2D`)

kmplot(`TEXT-SOFT(2)_2D`)
}
\keyword{datasets}
