% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TH3RESA_2A}
\alias{TH3RESA_2A}
\title{TH3RESA, figure 2A}
\format{
A data frame of 602 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (physicians_choice, trastuzumab_emtansine) \cr
}
}
\source{
Krop IE, Kim S-B, González-Martín A, et al. Trastuzumab emtansine
versus treatment of physician’s choice for pretreated HER2-positive
advanced breast cancer (TH3RESA): a randomised, open-label, phase 3
trial. Lancet Oncol 2014; 15: 689–99.
}
\usage{
TH3RESA_2A
}
\description{
Kaplan-Meier digitized data from TH3RESA, figure 2A (PMID 24793816). A reported sample size of 602 for a primary endpoint of PFS/OS in breast cancer.
}
\examples{
summary(TH3RESA_2A)

kmplot(TH3RESA_2A)
}
\keyword{datasets}
