% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRAPEZE_1A}
\alias{TRAPEZE_1A}
\title{TRAPEZE, figure 1A}
\format{
A data frame of 757 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (no_za, za) \cr
}
}
\source{
James ND, Pirrie SJ, Pope AM, et al. Clinical Outcomes and Survival
Following Treatment of Metastatic Castrate-Refractory Prostate Cancer
With Docetaxel Alone or With Strontium-89, Zoledronic Acid, or Both:
The TRAPEZE Randomized Clinical Trial. JAMA Oncol 2016; 2: 493–9.
}
\usage{
TRAPEZE_1A
}
\description{
Kaplan-Meier digitized data from TRAPEZE, figure 1A (PMID 26794729). A reported sample size of 757 for a primary endpoint of cPFS,cost-effectiveness in prostate cancer.
}
\examples{
summary(TRAPEZE_1A)

kmplot(TRAPEZE_1A)
}
\keyword{datasets}
