% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRIBE(2)_2B}
\alias{TRIBE(2)_2B}
\title{TRIBE(2), figure 2B}
\format{
A data frame of 508 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (folfiri_bevacizumab, folfoxiri_bevacizumab) \cr
}
}
\source{
Loupakis F, Cremolini C, Masi G, et al. Initial Therapy with
FOLFOXIRI and Bevacizumab for Metastatic Colorectal Cancer. N Engl J
Med 2014; 371: 1609-18.
}
\usage{
`TRIBE(2)_2B`
}
\description{
Kaplan-Meier digitized data from TRIBE(2), figure 2B (PMID 25337750). A reported sample size of 508 for a primary endpoint of PFS in colorectal cancer.
}
\examples{
summary(`TRIBE(2)_2B`)

kmplot(`TRIBE(2)_2B`)
}
\keyword{datasets}
