% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRIBE_2}
\alias{TRIBE_2}
\title{TRIBE, figure 2}
\format{
A data frame of 508 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (folfiri_bev, folfoxiri_bev) \cr
}
}
\source{
Cremolini C, Loupakis F, Antoniotti C, et al. FOLFOXIRI plus
bevacizumab versus FOLFIRI plus bevacizumab as first-line treatment
of patients with metastatic colorectal cancer: updated overall
survival and molecular subgroup analyses of the open-label, phase 3
TRIBE study. Lancet Oncol 2015; 16: 1306–15.
}
\usage{
TRIBE_2
}
\description{
Kaplan-Meier digitized data from TRIBE, figure 2 (PMID 26338525). A reported sample size of 508 for a primary endpoint of PFS in colorectal cancer.
}
\examples{
summary(TRIBE_2)

kmplot(TRIBE_2)
}
\keyword{datasets}
