% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TROG0306_2}
\alias{TROG0306_2}
\title{TROG0306, figure 2}
\format{
A data frame of 293 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in years) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (delayed_adt, immediate_adt) \cr
}
}
\source{
Duchesne GM, Woo HH, Bassett JK, et al. Timing of
androgen-deprivation therapy in patients with prostate cancer with a
rising PSA (TROG 03.06 and VCOG PR 01-03 [TOAD]): a randomised,
multicentre, non-blinded, phase 3 trial. Lancet Oncol 2016; 17:
727–37.
}
\usage{
TROG0306_2
}
\description{
Kaplan-Meier digitized data from TROG0306, figure 2 (PMID 27155740). A reported sample size of 293 for a primary endpoint of OS in prostate cancer.
}
\examples{
summary(TROG0306_2)

kmplot(TROG0306_2)
}
\keyword{datasets}
