% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TURANDOT_2B}
\alias{TURANDOT_2B}
\title{TURANDOT, figure 2B}
\format{
A data frame of 564 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab OS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (bev_capecitabine, bev_paclitaxel) \cr
}
}
\source{
Zielinski C, Láng I, Inbar M, et al. Bevacizumab plus paclitaxel
versus bevacizumab plus capecitabine as first-line treatment for
HER2-negative metastatic breast cancer (TURANDOT): primary endpoint
results of a randomised, open-label, non-inferiority, phase 3 trial.
Lancet Oncol 2016; 17: 1230–9.
}
\usage{
TURANDOT_2B
}
\description{
Kaplan-Meier digitized data from TURANDOT, figure 2B (PMID 27501767). A reported sample size of 564 for a primary endpoint of OS in breast cancer.
}
\examples{
summary(TURANDOT_2B)

kmplot(TURANDOT_2B)
}
\keyword{datasets}
