% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VANTAGE014_4}
\alias{VANTAGE014_4}
\title{VANTAGE014, figure 4}
\format{
A data frame of 661 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in weeks) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (placebo, vorinostat) \cr
}
}
\source{
Krug LM, Kindler HL, Calvert H, et al. Vorinostat in patients with
advanced malignant pleural mesothelioma who have progressed on
previous chemotherapy (VANTAGE-014): a phase 3, double-blind,
randomised, placebo-controlled trial. Lancet Oncol 2015; 16: 447–56.
}
\usage{
VANTAGE014_4
}
\description{
Kaplan-Meier digitized data from VANTAGE014, figure 4 (PMID 25800891). A reported sample size of 661 for a primary endpoint of OS/safety in lung cancer.
}
\examples{
summary(VANTAGE014_4)

kmplot(VANTAGE014_4)
}
\keyword{datasets}
