% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WJOG5108L_2A}
\alias{WJOG5108L_2A}
\title{WJOG5108L, figure 2A}
\format{
A data frame of 559 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (erlotinib, gefitinib) \cr
}
}
\source{
Urata Y, Katakami N, Morita S, et al. Randomized Phase III Study
Comparing Gefitinib With Erlotinib in Patients With Previously
Treated Advanced Lung Adenocarcinoma: WJOG 5108L. J Clin Oncol 2016;
34: 3248–57.
}
\usage{
WJOG5108L_2A
}
\description{
Kaplan-Meier digitized data from WJOG5108L, figure 2A (PMID 27022112). A reported sample size of 561 for a primary endpoint of PFS in lung cancer.
}
\examples{
summary(WJOG5108L_2A)

kmplot(WJOG5108L_2A)
}
\keyword{datasets}
