% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WJOG5208L_2B}
\alias{WJOG5208L_2B}
\title{WJOG5208L, figure 2B}
\format{
A data frame of 349 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (cisplatin_docetaxel, nedaplatin_docetaxel) \cr
}
}
\source{
Shukuya T, Yamanaka T, Seto T, et al. Nedaplatin plus docetaxel
versus cisplatin plus docetaxel for advanced or relapsed squamous
cell carcinoma of the lung (WJOG5208L): a randomised, open-label,
phase 3 trial. Lancet Oncol 2015; 16: 1630–8.
}
\usage{
WJOG5208L_2B
}
\description{
Kaplan-Meier digitized data from WJOG5208L, figure 2B (PMID 26522337). A reported sample size of 355 for a primary endpoint of OS in lung cancer.
}
\examples{
summary(WJOG5208L_2B)

kmplot(WJOG5208L_2B)
}
\keyword{datasets}
