% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WSG-ARA_3}
\alias{WSG-ARA_3}
\title{WSG-ARA, figure 3}
\format{
A data frame of 1,170 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab EFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (da_negative, da_positive) \cr
}
}
\source{
Nitz U, Gluz O, Zuna I, et al. Final results from the prospective
phase III WSG-ARA trial: impact of adjuvant darbepoetin alfa on
event-free survival in early breast cancer. Ann Oncol 2014; 25:
75–80.
}
\usage{
`WSG-ARA_3`
}
\description{
Kaplan-Meier digitized data from WSG-ARA, figure 3 (PMID 24356620). A reported sample size of 1,234 for a primary endpoint of EFS in breast cancer.
}
\examples{
summary(`WSG-ARA_3`)

kmplot(`WSG-ARA_3`)
}
\keyword{datasets}
