% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WSGAGO_2A}
\alias{WSGAGO_2A}
\title{WSGAGO, figure 2A}
\format{
A data frame of 1,930 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab EFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (control, ed_doc) \cr
}
}
\source{
Nitz U, Gluz O, Huober J, et al. Final analysis of the prospective
WSG-AGO EC-Doc versus FEC phase III trial in intermediate-risk (pN1)
early breast cancer: efficacy and predictive value of Ki67
expression. Ann Oncol 2014; 25: 1551–7.
}
\usage{
WSGAGO_2A
}
\description{
Kaplan-Meier digitized data from WSGAGO, figure 2A (PMID 24827128). A reported sample size of 2,011 for a primary endpoint of EFS in breast cancer.
}
\examples{
summary(WSGAGO_2A)

kmplot(WSGAGO_2A)
}
\keyword{datasets}
