% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X20050181_1B}
\alias{X20050181_1B}
\title{20050181, figure 1B}
\format{
A data frame of 486 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab PFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (folfiri, panitumumab_folfiri) \cr
}
}
\source{
Peeters M, Price TJ, Cervantes A, et al. Final results from a
randomized phase 3 study of FOLFIRI {+/-} panitumumab for second-line
treatment of metastatic colorectal cancer. Ann Oncol 2014; 25:
107–16.
}
\usage{
X20050181_1B
}
\description{
Kaplan-Meier digitized data from 20050181, figure 1B (PMID 24356622). A reported sample size of 1,186 for a primary endpoint of PFS/OS in colorectal cancer.
}
\examples{
summary(X20050181_1B)

kmplot(X20050181_1B)
}
\keyword{datasets}
