% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X2005B3030_2A}
\alias{X2005B3030_2A}
\title{2005B3030, figure 2A}
\format{
A data frame of 156 observations and 3 variables:
\tabular{lll}{
\tab \code{time} \tab event time (in months) \cr
\tab \code{event} \tab DFS event indicator (\code{0}: no event, \code{1}: event) \cr
\tab \code{arm} \tab treatment arms (control, pci) \cr
}
}
\source{
Li N, Zeng Z-F, Wang S-Y, et al. Randomized phase III trial of
prophylactic cranial irradiation versus observation in patients with
fully resected stage IIIA-N2 nonsmall-cell lung cancer and high risk
of cerebral metastases after adjuvant chemotherapy. Ann Oncol 2015;
26: 504–9.
}
\usage{
X2005B3030_2A
}
\description{
Kaplan-Meier digitized data from 2005B3030, figure 2A (PMID 25515658). A reported sample size of 156 for a primary endpoint of DFS in lung cancer.
}
\examples{
summary(X2005B3030_2A)

kmplot(X2005B3030_2A)
}
\keyword{datasets}
